% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metric_observer.R
\name{add_metric_observer}
\alias{add_metric_observer}
\title{Add Metric Observer}
\usage{
add_metric_observer(snapshot, metric)
}
\arguments{
\item{snapshot}{The snapshot \code{list} object that is passed on from \code{run_app()}}

\item{metric}{A reactive value of the metric input found in the \code{ui} function.}
}
\description{
Add Metric Observer
}
