% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{applyPrettyDomainNames}
\alias{applyPrettyDomainNames}
\title{Apply user-facing domain names}
\usage{
applyPrettyDomainNames(lDomains)
}
\arguments{
\item{lDomains}{\code{list} Named list of data domain data.frames.}
}
\value{
The list of domain dfs, with better user-facing names.
}
\description{
Apply user-facing domain names
}
\keyword{internal}
