% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AsFunction}
\alias{AsFunction}
\title{Load a Function for a Plugin}
\usage{
AsFunction(strFunction)
}
\arguments{
\item{strFunction}{\code{character} The name of a function, with or without
namespace ("reactive" or "shiny::reactive"). If a namespace is supplied, we
check that the package is installed, and attach it. The function can also
be supplied as a function or a purrr-style formula. See
\code{\link[rlang:as_function]{rlang::as_function()}} (the \code{x} argument) for more details.}
}
\value{
The function, if it can be found.
}
\description{
Load a Function for a Plugin
}
\keyword{internal}
