% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CallIfReactive}
\alias{CallIfReactive}
\title{Resolve a potential reactive object}
\usage{
CallIfReactive(x)
}
\arguments{
\item{x}{An object that might be reactive.}
}
\value{
Either \code{x} (if \code{x} isn't reactive) or \code{x()}.
}
\description{
Resolve a potential reactive object
}
\keyword{internal}
