% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{CheckIsIn}
\alias{CheckIsIn}
\title{Check that a vector is in another vector}
\usage{
CheckIsIn(x, chrAllowedValues, strWhat = rlang::caller_arg(x), envCall)
}
\arguments{
\item{x}{An object to validate.}

\item{chrAllowedValues}{\code{character} A vector of allowed values for a vector.}

\item{strWhat}{\code{character} A sentence-case description of the object being
inspected.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated object.
}
\description{
Check that a vector is in another vector
}
\keyword{internal}
