% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colorScheme.R
\name{colorScheme}
\alias{colorScheme}
\title{gsm Color Scheme}
\usage{
colorScheme(
  strColorName = c("gray", "green", "amber", "red"),
  strColorFamily = c("dark", "light")
)
}
\arguments{
\item{strColorName}{\code{character} The target color to report about.}

\item{strColorFamily}{\code{character} Whether to load the \code{"dark"} version of
this color, or the \code{"light"} version. Default: \code{"dark"}.}
}
\value{
The hex code of a color.
}
\description{
Eventually this should have a single home, but it doesn't make sense yet for
anything to export it, so it exists in both gsm and gsm.app.
}
\keyword{internal}
