% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CombineDomainData.R
\name{CombineDomainData}
\alias{CombineDomainData}
\title{Combine domain data}
\usage{
CombineDomainData(dfDomain_Study, dfDomain_Group, dfDomain_Participant)
}
\arguments{
\item{dfDomain_Study}{\code{data.frame} Detailed data about a specific domain,
such as "Adverse Events", at the study level.}

\item{dfDomain_Group}{\code{data.frame} Detailed data about a specific domain,
such as "Adverse Events", at the group level.}

\item{dfDomain_Participant}{\code{data.frame} Detailed data about a specific
domain, such as "Adverse Events", at the participant level.}
}
\value{
The combined data, with the \code{VizLevel} column as a factor.
}
\description{
Combine domain data
}
\keyword{internal}
