% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\name{ConstructDataCounter}
\alias{ConstructDataCounter}
\title{Construct a Data Count Function}
\usage{
ConstructDataCounter(fnFetchData)
}
\arguments{
\item{fnFetchData}{\code{function} A function that takes a \code{strDomainID} argument
and optional \code{strGroupID}, \code{strGroupLevel}, \code{strSubjectID}, and/or
\code{dSnapshotDate}, and returns a data.frame. See \code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for
an example. The returned data.frame contains information about the named
domain. If the function throws an error, the error is elevated to the user,
so you can use errors to pass requirements through to the user.}
}
\value{
A function that returns counts of rows in the specified domain data.
}
\description{
It is often faster to fetch \emph{counts} of data, rather than the entire dataset.
If your \code{fnFetchData()} function is already fast, however, this function
allows you to construct a simple data counter from your data fetcher.
}
\examples{
fnDataCounter <- ConstructDataCounter(sample_fnFetchData)
fnDataCounter("AE") == nrow(sample_fnFetchData("AE"))
}
