% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_server.R
\name{div_ParticipantCard_Placeholder}
\alias{div_ParticipantCard_Placeholder}
\title{The placeholder portion of a participant card}
\usage{
div_ParticipantCard_Placeholder()
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} asking the user to select a participant.
}
\description{
The placeholder portion of a participant card
}
\keyword{internal}
