% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{extract_field_name}
\alias{extract_field_name}
\title{Get a field name from a variable name}
\usage{
extract_field_name(strArgName)
}
\arguments{
\item{strArgName}{The name of the argument.}
}
\value{
\code{strArgName} with the beginning part removed.
}
\description{
Assume standard argument name format, like \code{strGroupID} or \code{dSnapshotDat}.
}
\keyword{internal}
