% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ExtractAppTitle}
\alias{ExtractAppTitle}
\title{Extract an App Title from dfGroups}
\usage{
ExtractAppTitle(dfGroups)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}
}
\value{
A length-1 character vector with the first available of "nickname",
"protocol_title", or "studyid", if any are available as a \code{Param} in
\code{dfGroups} with \code{GroupLevel == "Study"}, truncated to at most 30
characters. If none of those are available, \code{"GSM Deep Dive"}.
}
\description{
Extract the "nickname", "protocol_title", or "studyid" of the study from
\code{dfGroup} to use as a title for the app.
}
\examples{
ExtractAppTitle(sample_dfGroups)
ExtractAppTitle(
  data.frame(
    GroupLevel = "Study",
    Param = "nickname",
    Value = "This is a very long nickname which will be truncated"
  )
)
}
