% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{ExtractFieldName}
\alias{ExtractFieldName}
\title{Get a field name from a variable name}
\usage{
ExtractFieldName(strArgName)
}
\arguments{
\item{strArgName}{The name of the argument.}
}
\value{
\code{strArgName} with the beginning part removed.
}
\description{
Assume standard argument name format, like \code{strGroupID} or \code{dSnapshotDate}.
}
\keyword{internal}
