% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{filter_byGroupID}
\alias{filter_byGroupID}
\title{Filter by GroupID}
\usage{
filter_byGroupID(df, strGroupID)
}
\arguments{
\item{df}{A data.frame to filter.}

\item{strGroupID}{\code{character} A \code{GroupID} to focus on.}
}
\value{
The filtered data.frame.
}
\description{
Filter by GroupID
}
\keyword{internal}
