% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{FilterBefore}
\alias{FilterBefore}
\title{Filter to only values before a date}
\usage{
FilterBefore(df, strField, dMaxDate)
}
\arguments{
\item{df}{A data.frame to filter.}

\item{strField}{The name of the field.}

\item{dMaxDate}{The date that values should be before.}
}
\value{
The filtered df.
}
\description{
Filter to only values before a date
}
\keyword{internal}
