% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{FilterByLatestIfPresent}
\alias{FilterByLatestIfPresent}
\title{Filter to latest snapshot if field exists}
\usage{
FilterByLatestIfPresent(df)
}
\arguments{
\item{df}{A data.frame to filter.}
}
\value{
The filtered df.
}
\description{
Filter to latest snapshot if field exists
}
\keyword{internal}
