% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{FilterbyMetricID}
\alias{FilterbyMetricID}
\title{Filter by MetricID}
\usage{
FilterbyMetricID(df, strMetricID)
}
\arguments{
\item{df}{A data.frame to filter.}

\item{strMetricID}{\code{character} A \code{MetricID} to focus on.}
}
\value{
The filtered data.frame.
}
\description{
Filter by MetricID
}
\keyword{internal}
