% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dfDomains.R
\name{FilterDomainData}
\alias{FilterDomainData}
\title{Filter Domain Data by Relevant Fields}
\usage{
FilterDomainData(
  df,
  strDomainID,
  dSnapshotDate = NULL,
  dfGroups = NULL,
  strGroupLevel = NULL,
  strGroupID = NULL,
  strSubjectID = NULL
)
}
\arguments{
\item{df}{\code{data.frame} A data frame to wrangle.}

\item{strDomainID}{\code{character} The domain data.frame to load.}

\item{dSnapshotDate}{\code{Date} The date of a data snapshot.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: \code{NULL}
and taken from \code{dfMetrics$GroupLevel} if available.}

\item{strGroupID}{\code{character} A \code{GroupID} of an individual site or other
group within a study.}

\item{strSubjectID}{\code{character} A \code{SubjectID} of an individual participant.}
}
\value{
The data.frame, with filters applied.
}
\description{
Filter Domain Data by Relevant Fields
}
\keyword{internal}
