% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{FinalWorkflowOutputName}
\alias{FinalWorkflowOutputName}
\title{Find the Name of the Final Step}
\usage{
FinalWorkflowOutputName(lWorkflow)
}
\value{
A string with the "output" name of the final step of a workflow.
}
\description{
Find the Name of the Final Step
}
\keyword{internal}
