% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{FindCategoricalFieldNames}
\alias{FindCategoricalFieldNames}
\title{Find categorical fields in a data frame}
\usage{
FindCategoricalFieldNames(
  df,
  chrExcludes = c("SubjectID", "GroupID", "GroupLevel")
)
}
\arguments{
\item{df}{\code{data.frame} A data frame to wrangle.}

\item{chrExcludes}{\code{character} Terms to \emph{not} include in the returned vector.}
}
\value{
A character vector of categorical-like field names. Currently this
include all characters and factors for which at least 1 value is repeated.
}
\description{
Find categorical fields in a data frame
}
\keyword{internal}
