% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{FindNonZeroDecimals}
\alias{FindNonZeroDecimals}
\title{Find interesting digits for rounding}
\usage{
FindNonZeroDecimals(dblX, intMaxDecimals = 5L)
}
\arguments{
\item{dblX}{A vector of numbers (usually with decimal places).}

\item{intMaxDecimals}{The maximum number of decimal places to keep.}
}
\value{
An integer between \code{0} and \code{intMaxDecimals}, representing the number
of decimal places to keep.
}
\description{
Find interesting digits for rounding
}
\keyword{internal}
