% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_from_yaml.R
\name{get_metadata_from_yaml}
\alias{get_metadata_from_yaml}
\title{Ingest metadata}
\usage{
get_metadata_from_yaml(
  type = NULL,
  path = system.file(type, package = "gsmApp")
)
}
\arguments{
\item{type}{\code{character} type of metadata
\itemize{
\item assessments
\item workflows
}}

\item{path}{\code{character} location of .yaml files containing metadata}
}
\value{
\code{list} list of assessment metadata
}
\description{
Ingest metadata
}
