% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conditions.R
\name{gsmapp_abort}
\alias{gsmapp_abort}
\title{Error messaging for this package.}
\usage{
gsmapp_abort(message, class, ..., .envir = rlang::caller_env())
}
\arguments{
\item{message}{It is formatted via a call to \code{\link[cli:cli_bullets]{cli_bullets()}}.}

\item{class}{A descriptive label for this type of error, in lower_snake_case.}

\item{...}{Passed to \code{\link[rlang:abort]{rlang::abort()}}, \code{\link[rlang:abort]{rlang::warn()}} or
\code{\link[rlang:abort]{rlang::inform()}}.}

\item{.envir}{Environment in which any variables used in the message are
defined. You almost definitely want to leave this as the default value.}
}
\value{
An error condition with classes \code{"gsm.app-condition"},
\code{"gsm.app-error"}, and \code{"gsm.app-error-{class}"}.
}
\description{
Error messaging for this package.
}
\keyword{internal}
