% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp_Server.R
\name{gsmApp_Server}
\alias{gsmApp_Server}
\title{Define Server}
\usage{
gsmApp_Server(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  fnFetchData,
  fnCountData = ConstructDataCounter(fnFetchData),
  chrDomains = c(AE = "Adverse Events", DATACHG = "Data Changes", DATAENT = "Data Entry",
    ENROLL = "Enrollment", LB = "Lab", PD = "Protocol Deviations", QUERY = "Queries",
    STUDCOMP = "Study Completion", SUBJ = "Subject Metadata", SDRGCOMP =
    "Treatment Completion"),
  lPlugins = NULL,
  fnServer = NULL
)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data. This
data.frame is created by binding together the \code{lAnalysis} "Analysis_Input"
list of data.frames output of analysis workflows, with a \code{MetricID} column
for the name of each list.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created with \code{\link[gsm.reporting:MakeBounds]{gsm.reporting::MakeBounds()}}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created with \code{\link[gsm.reporting:MakeMetric]{gsm.reporting::MakeMetric()}}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created with \code{\link[gsm.reporting:BindResults]{gsm.reporting::BindResults()}}.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomainID} argument
and optional \code{strGroupID}, \code{strGroupLevel}, \code{strSubjectID}, and/or
\code{dSnapshotDate}, and returns a data.frame. See \code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for
an example. The returned data.frame contains information about the named
domain. If the function throws an error, the error is elevated to the user,
so you can use errors to pass requirements through to the user.}

\item{fnCountData}{\code{function} A function that takes a \code{strDomainID} argument
and optional \code{strGroupID}, \code{strGroupLevel}, \code{strSubjectID}, and/or
\code{dSnapshotDate}, and returns an integer count of the number of rows in that
filtered domain. By default, this function is constructed using
\code{fnFetchData()} and \code{\link[=ConstructDataCounter]{ConstructDataCounter()}}.}

\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{fnServer}{\code{function} A Shiny server function that takes arguments
\code{input}, \code{output}, and \code{session}. This function will be called at the start
of the main app server function.}
}
\value{
The main server function for use in a shiny app.
}
\description{
Define Server
}
\keyword{internal}
