% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_Plot.R
\name{gsmWidgetOutput}
\alias{gsmWidgetOutput}
\title{Pick Up Dependencies for Rendered Output}
\usage{
gsmWidgetOutput(
  strWidgetName,
  outputId,
  width = "100\%",
  height = "400px",
  excludes = character(),
  ...
)
}
\arguments{
\item{strWidgetName}{\code{character} The name of a widget in the gsm package.}

\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{excludes}{An optional vector of named dependencies to \emph{not} include.}

\item{...}{
  Arguments passed on to \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput}}
  \describe{
    \item{\code{inline}}{use an inline (\code{span()}) or block container (\code{div()})
for the output}
    \item{\code{reportSize}}{Should the widget's container size be reported in the
shiny session's client data?}
    \item{\code{fill}}{whether or not the returned tag should be treated as a fill item,
meaning that its \code{height} is allowed to grow/shrink to fit a fill container
with an opinionated height (see \code{\link[htmltools:bindFillRole]{htmltools::bindFillRole()}} for more).
Examples of fill containers include \code{bslib::card()} and
\code{bslib::card_body_fill()}.}
  }}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} with the \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}
and its dependencies.
}
\description{
Pick Up Dependencies for Rendered Output
}
\keyword{internal}
