% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable_UI.R
\name{gtIO}
\alias{gtIO}
\title{gt Table Input and Output}
\usage{
gtIO(id)
}
\arguments{
\item{id}{}
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} containing the dependencies needed to use
gt as both an input and an output, and a \code{\link[shiny:htmlOutput]{shiny::htmlOutput()}} with class
"gtIO".
}
\description{
gt Table Input and Output
}
\keyword{internal}
