% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BarChart.R
\name{htmlDependency_BarChartInput}
\alias{htmlDependency_BarChartInput}
\title{Bar Chart Input JavaScript}
\usage{
htmlDependency_BarChartInput()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless how many times they are added.
}
\description{
Attach \code{barChartInput.js} to an app or other HTML exactly once.
}
\keyword{internal}
