% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_DetectCardClicks.R
\name{htmlDependency_DetectCardClicks}
\alias{htmlDependency_DetectCardClicks}
\title{DetectCardClicks JavaScript}
\usage{
htmlDependency_DetectCardClicks()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{detectCardClicks.js} to an app or other HTML exactly once.
}
\keyword{internal}
