% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupOverview.R
\name{htmlDependency_GroupOverview}
\alias{htmlDependency_GroupOverview}
\title{GroupOverview Dependencies}
\usage{
htmlDependency_GroupOverview()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless how many times they are added.
}
\description{
GroupOverview Dependencies
}
\keyword{internal}
