% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable_UI.R
\name{htmlDependency_gtIO}
\alias{htmlDependency_gtIO}
\title{gt Input-Output Dependencies}
\usage{
htmlDependency_gtIO()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless how many times they are added.
}
\description{
Attach CSS and JavaScript necessary for gtIO to an app or other HTML exactly
once.
}
\keyword{internal}
