% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable_UI.R
\name{htmlDependency_gtIOInput}
\alias{htmlDependency_gtIOInput}
\title{gt Input JavaScript}
\usage{
htmlDependency_gtIOInput()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{gtIOInput.js} to an app or other HTML exactly once.
}
\keyword{internal}
