% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable_UI.R
\name{htmlDependency_gtIOjs}
\alias{htmlDependency_gtIOjs}
\title{gt JavaScript}
\usage{
htmlDependency_gtIOjs()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{gtIO.js} to an app or other HTML exactly once.
}
\keyword{internal}
