% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlDependency.R
\name{htmlDependency_HighlightTableRow}
\alias{htmlDependency_HighlightTableRow}
\title{HighlightTableRow javascript}
\usage{
htmlDependency_HighlightTableRow()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{HighlightTableRow.js} to an app or other HTML exactly once.
}
\keyword{internal}
