% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_NavbarExtras.R
\name{htmlDependency_NavbarExtras}
\alias{htmlDependency_NavbarExtras}
\title{NavbarExtras Dependencies}
\usage{
htmlDependency_NavbarExtras()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless of how many times they are added.
}
\description{
NavbarExtras Dependencies
}
\keyword{internal}
