% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_NavbarExtras.R
\name{htmlDependency_NavbarExtrasPosition}
\alias{htmlDependency_NavbarExtrasPosition}
\title{NavbarExtrasPosition JavaScript}
\usage{
htmlDependency_NavbarExtrasPosition()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless of how many
times it is added.
}
\description{
Attach \code{navbarExtrasPosition.js} to an app or other HTML exactly once.
}
\keyword{internal}
