% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlDependency.R
\name{htmlDependency_Stylesheet}
\alias{htmlDependency_Stylesheet}
\title{gsm.app stylesheet}
\usage{
htmlDependency_Stylesheet(
  filename,
  name = sub("\\\\.css$", "", filename),
  version = "1.0.0"
)
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach a stylesheet from this package to an app or other HTML exactly once.
}
\keyword{internal}
