% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlDependency.R
\name{htmlDependency_TableClick}
\alias{htmlDependency_TableClick}
\title{TableClick javascript}
\usage{
htmlDependency_TableClick()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{tableClick.js} to an app or other HTML exactly once.
}
\keyword{internal}
