% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TogglePill.R
\name{htmlDependency_TogglePill}
\alias{htmlDependency_TogglePill}
\title{TogglePill Dependencies}
\usage{
htmlDependency_TogglePill()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless of how many times they are added.
}
\description{
Attach CSS and JavaScript necessary for TogglePill to an app or other HTML
exactly once.
}
\keyword{internal}
