% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TogglePill_UI.R
\name{htmlDependency_updateStyleJS}
\alias{htmlDependency_updateStyleJS}
\title{Style JavaScript}
\usage{
htmlDependency_updateStyleJS()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{updateStyle.js} to an app or other HTML exactly once.
}
\keyword{internal}
