% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_WidgetPlot_UI.R
\name{htmlDependency_WidgetPlotInput}
\alias{htmlDependency_WidgetPlotInput}
\title{Widget Plot Input JavaScript}
\usage{
htmlDependency_WidgetPlotInput(strWidgetName)
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless how many times they are added.
}
\description{
Attach the appropriate \verb{*Input.js} to an app or other HTML exactly once.
}
\keyword{internal}
