% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlot_UI.R
\name{htmlDependency_WidgetPlotInputHelpers}
\alias{htmlDependency_WidgetPlotInputHelpers}
\title{Widget Plot Input javascript}
\usage{
htmlDependency_WidgetPlotInputHelpers()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{widgetPlotInputHelpers.js} to an app or other HTML exactly once.
}
\keyword{internal}
