% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_metric_select.R
\name{initialize_metric_select}
\alias{initialize_metric_select}
\title{Initialize Metric Select}
\usage{
initialize_metric_select(input, output, session, snapshot)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}
}
\description{
Initialize Metric Select
}
