% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_site_select.R
\name{initialize_site_select}
\alias{initialize_site_select}
\title{Initialize Site Select}
\usage{
initialize_site_select(input, output, session, snapshot)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Initialize Site Select
}
