% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_numeric.R
\name{input_numeric}
\alias{input_numeric}
\title{Generates nuemric input}
\usage{
input_numeric(id, default)
}
\arguments{
\item{id}{\code{character} Input ID}

\item{default}{\code{numeric} Input metadata}
}
\description{
Generates nuemric input
}
