% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_assessment_server.R
\name{make_assessment_server}
\alias{make_assessment_server}
\title{Define a Shiny server function given an assessment}
\usage{
make_assessment_server(assessment)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny server function
}
\description{
Define a Shiny server function given an assessment
}
