% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_assessment_ui.R
\name{make_assessment_ui}
\alias{make_assessment_ui}
\title{Define a Shiny UI function given an assessment}
\usage{
make_assessment_ui(assessment)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny UI function
}
\description{
Define a Shiny UI function given an assessment
}
