% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{make_dfParticipantGroups}
\alias{make_dfParticipantGroups}
\title{Filter and sort participants by group}
\usage{
make_dfParticipantGroups(dfAnalyticsInput)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}
}
\value{
A data.frame with \code{SubjectID} and \code{GroupID}.
}
\description{
Filter and sort participants by group
}
\keyword{internal}
