% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{make_lStudy}
\alias{make_lStudy}
\title{Reshape group and result info into study information}
\usage{
make_lStudy(dfGroups, dfResults)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}
}
\value{
A list with study information.
}
\description{
Reshape group and result info into study information
}
\keyword{internal}
