% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_overview_table_server.R
\name{make_overview_table_server}
\alias{make_overview_table_server}
\title{Define a Shiny server function given an assessment}
\usage{
make_overview_table_server(workflows)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny server function
}
\description{
Define a Shiny server function given an assessment
}
