% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_overview_table_ui.R
\name{make_overview_table_ui}
\alias{make_overview_table_ui}
\title{Define a Shiny UI function given an assessment}
\usage{
make_overview_table_ui(workflows)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny UI function
}
\description{
Define a Shiny UI function given an assessment
}
