% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_make_summary_table.R
\name{make_summary_table}
\alias{make_summary_table}
\title{Create Summary table in KRIReport.Rmd for each KRI}
\usage{
make_summary_table(lAssessment, dfSite = NULL)
}
\arguments{
\item{lAssessment}{\code{list} List of KRI assessments from \code{params} within
\code{KRIReport.Rmd}.}

\item{dfSite}{\code{data.frame} Optional site-level metadata.}
}
\description{
Create Summary table in KRIReport.Rmd for each KRI
}
\keyword{internal}
