% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_workflow_server.R
\name{make_workflow_server}
\alias{make_workflow_server}
\title{Define a Shiny server function given an workflow}
\usage{
make_workflow_server(workflow, assessments, thresholds, domain_data)
}
\arguments{
\item{workflow}{\code{list} workflow specification}
}
\value{
\code{function} Shiny server function
}
\description{
Define a Shiny server function given an workflow
}
