% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_workflow_ui.R
\name{make_workflow_ui}
\alias{make_workflow_ui}
\title{Define a Shiny UI function given an workflow}
\usage{
make_workflow_ui(workflow)
}
\arguments{
\item{workflow}{\code{list} workflow specification}
}
\value{
\code{function} Shiny UI function
}
\description{
Define a Shiny UI function given an workflow
}
