% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugin.R
\name{MakePrefixRegex}
\alias{MakePrefixRegex}
\title{Construct Regex to Find Prefixes}
\usage{
MakePrefixRegex(
  chrPrefixes = c("chr", "df", "d", "fct", "int", "l", "num", "rctv", "str")
)
}
\arguments{
\item{chrPrefixes}{\code{character} Argument "prefixes" (really can appear
anywhere in the argument name, as long as they're preceded by either the
beginning of the name or a non-alphanumeric character, and followed by a
capital letter or a non-alphanumeric character).}
}
\value{
A regex to find the prefixes (to remove them).
}
\description{
Construct Regex to Find Prefixes
}
\keyword{internal}
