% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_meta_to_data_frame.R
\name{map_meta_to_data_frame}
\alias{map_meta_to_data_frame}
\title{Structure {gsm} Metadata}
\usage{
map_meta_to_data_frame(
  meta_in = c(yaml::read_yaml(system.file("mappings", "mapping_rawplus.yaml", package =
    "gsm")), yaml::read_yaml(system.file("mappings", "mapping_edc.yaml", package =
    "gsm")))
)
}
\arguments{
\item{meta_in}{\code{list} {gsm} domain/column/value mapping metadata with structure:
domain:
key: value}
}
\value{
\code{data.frame} {gsm} metadata in tabular form expected by {safetyGraphics}
}
\description{
Map {gsm} metadata to the tabular structure expected by {safetyGraphics}
}
\examples{
meta <- map_meta_to_data_frame()

}
