% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_meta_to_list.R
\name{map_meta_to_list}
\alias{map_meta_to_list}
\title{Structure {gsm} Metadata}
\usage{
map_meta_to_list(meta_in = gsmApp::meta_safetyGraphics)
}
\arguments{
\item{meta_in}{\code{data.frame} {safetyGraphics} metadata}
}
\value{
\code{list} metadata from {safetyGraphics} in list form expected by {gsm}
}
\description{
Map {safetyGraphics} metadata to the list structure expected by {gsm}
}
\examples{
meta_gsm <- map_meta_to_list()

}
