% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_meta_to_safetyGraphics.R
\name{map_meta_to_safetyGraphics}
\alias{map_meta_to_safetyGraphics}
\title{Structure {gsm} Metadata}
\usage{
map_meta_to_safetyGraphics(meta_in = clindata::mapping_rawplus)
}
\arguments{
\item{meta_in}{\code{list} {gsm} metadata}
}
\value{
\code{data.frame} {gsm} metadata in tabular form expected by {safetyGraphics}
}
\description{
Map {gsm} metadata to the tabular structure expected by {safetyGraphics}
}
\examples{
meta_safetyGraphics <- map_meta_to_safetyGraphics()

}
