% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metric_details_server.R
\name{metric_details_server}
\alias{metric_details_server}
\title{Metric Details Server}
\usage{
metric_details_server(id, snapshot, metric, site)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}

\item{metric}{The reactive value provided by the metric input from \code{server}}

\item{site}{The reactive value provided by the site input from \code{server}}
}
\description{
Metric Details Server
}
