% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metric_details_ui.R
\name{metric_details_ui}
\alias{metric_details_ui}
\title{Metric Details UI}
\usage{
metric_details_ui(id)
}
\arguments{
\item{id}{The namespace id for this module.}
}
\description{
Metric Details UI
}
