% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ActionList_Server.R
\name{mod_ActionList_Server}
\alias{mod_ActionList_Server}
\title{List of ActionLinks Server}
\usage{
mod_ActionList_Server(id, rctv_chrIDs)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_chrIDs}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a vector of ids for modules, inputs, or other elements.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} with the id of the most recently clicked
link.
}
\description{
List of ActionLinks Server
}
\keyword{internal}
