% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BarChart.R
\name{mod_BarChart_UI}
\alias{mod_BarChart_UI}
\title{BarChart Module UI}
\usage{
mod_BarChart_UI(id, strPlotTitle = NULL)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and a
\code{\link[=Widget_BarChartOutput]{Widget_BarChartOutput()}}.
}
\description{
BarChart Module UI
}
\keyword{internal}
