% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CascadingSelect.R
\name{mod_CascadingSelect_Server}
\alias{mod_CascadingSelect_Server}
\title{Cascading Select Input Server}
\usage{
mod_CascadingSelect_Server(
  id,
  rctv_strDependent,
  l_chrChoices,
  rctv_strSelection,
  lglUpdateLabel = FALSE
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strDependent}{\code{reactive} The reactive string value that this
module's selection depends on.}

\item{l_chrChoices}{\code{list} A named list where names correspond to values from
\code{rctv_strDependent} and values are the character vectors of choices.}

\item{rctv_strSelection}{\code{reactiveVal} A \code{reactiveVal} from the parent scope
that this module will update with its selection.}

\item{lglUpdateLabel}{\code{logical} A flag indicating whether to update the
select input's label with the value of \code{rctv_strDependent}.}
}
\value{
The server function does not return a value.
}
\description{
Cascading Select Input Server
}
\keyword{internal}
