% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CascadingSelect.R
\name{mod_CascadingSelect_UI}
\alias{mod_CascadingSelect_UI}
\title{Cascading Select Input UI}
\usage{
mod_CascadingSelect_UI(id, strLabel, chrChoices = "All")
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strLabel}{\code{character} The label of a field.}

\item{chrChoices}{\code{character} A (named, optionally) vector of choices for
this input.}
}
\value{
A \code{bslib::nav_item()}.
}
\description{
Cascading Select Input UI
}
\keyword{internal}
