% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CombineDomainData.R
\name{mod_CombineDomainData_Server}
\alias{mod_CombineDomainData_Server}
\title{Pseudo-module to cleanly combine domain dfs}
\usage{
mod_CombineDomainData_Server(
  id,
  l_rctvActive,
  rctv_dfDomain_Study,
  rctv_dfDomain_Group,
  rctv_dfDomain,
  rctv_strGroupLevel,
  rctv_strGroupID,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{l_rctvActive}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a populated character vector or something that becomes
\code{NULL} when passed through \code{\link[=NullifyEmpty]{NullifyEmpty()}}.}

\item{rctv_dfDomain_Study}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns a study-level domain dataframe.}

\item{rctv_dfDomain_Group}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns a group-level domain dataframe.}

\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{reactive} that returns the combined domain data.
}
\description{
Pseudo-module to cleanly combine domain dfs
}
\keyword{internal}
