% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataTableCard_Server.R
\name{mod_DataTableCard_Server}
\alias{mod_DataTableCard_Server}
\title{DataTable Card Server}
\usage{
mod_DataTableCard_Server(id, rctv_lData, rctv_strName)
}
\arguments{
\item{id}{The id for this module in this namespace.}

\item{rctv_lData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns a
named list of dataframes.}

\item{rctv_strName}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the name of an object, such as a particular dataframe in a named
list.}
}
\description{
DataTable Card Server
}
\keyword{internal}
