% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DataTableCard_UI.R
\name{mod_DataTableCard_UI}
\alias{mod_DataTableCard_UI}
\title{Card with DTOutput}
\usage{
mod_DataTableCard_UI(id)
}
\arguments{
\item{id}{The id for this module in this namespace.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with either a placeholder, or a \code{\link[DT:dataTableOutput]{DT::DTOutput()}}
table.
}
\description{
Card with DTOutput
}
\keyword{internal}
