% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainCount.R
\name{mod_DomainCount_Server}
\alias{mod_DomainCount_Server}
\title{Domain Tab Count Server}
\usage{
mod_DomainCount_Server(id, rctv_intDomainCount)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
Nothing. Called to update the output to the number.
}
\description{
Display a count of rows in a given domain.
}
\keyword{internal}
