% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainData.R
\name{mod_DomainData_UI}
\alias{mod_DomainData_UI}
\title{Domain Data Tab UI}
\usage{
mod_DomainData_UI(id, strDomainLabel, strDomainID)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strDomainLabel}{\code{character} The domain name to display. For example,
for ID \code{"AE"}, the label is \code{"Adverse Events"}. Can also be a
\code{\link[shiny:htmlOutput]{shiny::uiOutput()}} which renders to an inline string.}

\item{strDomainID}{\code{character} The domain data.frame to load.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} with either a placeholder, or a
\code{\link[gt:gt_output]{gt::gt_output()}} table.
}
\description{
Domain Data Tab UI
}
\keyword{internal}
