% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainDetails.R
\name{mod_DomainDetails_UI}
\alias{mod_DomainDetails_UI}
\title{Domain Details UI}
\usage{
mod_DomainDetails_UI(
  id,
  chrDomains = c(AE = "Adverse Events", DATACHG = "Data Changes", DATAENT = "Data Entry",
    ENROLL = "Enrollment", LB = "Lab", PD = "Protocol Deviations", QUERY = "Queries",
    STUDCOMP = "Study Completion", SUBJ = "Subject Metadata", SDRGCOMP =
    "Treatment Completion")
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}
}
\value{
A \code{\link[bslib:navset_underline]{bslib::navset_underline()}} with a tab for each available domain.
}
\description{
Domain Details UI
}
\keyword{internal}
