% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainSummary.R
\name{mod_DomainSummary_Server}
\alias{mod_DomainSummary_Server}
\title{Dynamically Generate the Domain Summary}
\usage{
mod_DomainSummary_Server(id, rctv_strDomainID, l_rctvDomains)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}

\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}
}
\description{
Dynamically Generate the Domain Summary
}
\keyword{internal}
