% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainSummary.R
\name{mod_DomainSummary_UI}
\alias{mod_DomainSummary_UI}
\title{Domain Summary UI}
\usage{
mod_DomainSummary_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with a \code{\link[=mod_ActionList_UI]{mod_ActionList_UI()}} showing domains and
row counts.
}
\description{
Domain Summary UI
}
\keyword{internal}
