% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DynamicLabelKey.R
\name{mod_DynamicLabelKey_Server}
\alias{mod_DynamicLabelKey_Server}
\title{Activate/inactivate buttons and return values}
\usage{
mod_DynamicLabelKey_Server(
  id,
  rctv_strGroupLevel,
  rctv_strGroupID,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A list of \code{\link[shiny:reactive]{shiny::reactive()}} objects to determine whether the given
button is selected.
}
\description{
Activate/inactivate buttons and return values
}
\keyword{internal}
