% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupOverview.R
\name{mod_GroupOverview_UI}
\alias{mod_GroupOverview_UI}
\title{Group Overview Module UI}
\usage{
mod_GroupOverview_UI(id, dfResults)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created with \code{\link[gsm.reporting:BindResults]{gsm.reporting::BindResults()}}.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with a \code{\link[=mod_RAGPillSet_UI]{mod_RAGPillSet_UI()}} and a corresponding
\code{\link[=Widget_GroupOverviewOutput]{Widget_GroupOverviewOutput()}}.
}
\description{
Group Overview Module UI
}
\keyword{internal}
