% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupParticipants.R
\name{mod_GroupParticipants_UI}
\alias{mod_GroupParticipants_UI}
\title{Group Participants UI}
\usage{
mod_GroupParticipants_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with a placeholder or a clickable table of group
participants.
}
\description{
Group Participants UI
}
\keyword{internal}
