% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gtBidirectional_Server.R
\name{mod_gtBidirectional_Server}
\alias{mod_gtBidirectional_Server}
\title{Bidirectional gt Table Selector Server}
\usage{
mod_gtBidirectional_Server(
  id,
  rctv_dfData,
  rctv_gtObject,
  rctv_strInput,
  strLabel,
  strEmpty = "None"
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_dfData}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a generic dataframe.}

\item{rctv_gtObject}{\verb{reactive gt_table} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a \code{\link[gt:gt]{gt::gt()}} object.}

\item{rctv_strInput}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the name of the target input.}

\item{strLabel}{\code{character} The label of a field.}

\item{strEmpty}{The value to return when everything is deselected.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} with the id(s) of the field(s) selected in
the table.
}
\description{
Bidirectional gt Table Selector Server
}
\keyword{internal}
