% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gtBidirectional_UI.R
\name{mod_gtBidirectional_UI}
\alias{mod_gtBidirectional_UI}
\title{Bidirectional gt Table Selector UI}
\usage{
mod_gtBidirectional_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[gt:gt_output]{gt::gt_output()}} that handles updates from external inputs.
}
\description{
Bidirectional gt Table Selector UI
}
\keyword{internal}
