% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_InlineSelect.R
\name{mod_InlineSelect_Server}
\alias{mod_InlineSelect_Server}
\title{Inline virtual select input server}
\usage{
mod_InlineSelect_Server(
  id,
  rctv_chrChoices = NULL,
  rctv_strLabel = NULL,
  rctv_strSelected = NULL,
  rctv_lglDisable = NULL,
  rctv_chrDisabledChoices = NULL,
  rctv_lglOpen = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_chrChoices}{\verb{reactive character} Choices to set.}

\item{rctv_strLabel}{\verb{reactive character} Label to set.}

\item{rctv_strSelected}{\verb{reactive character} Selected value to set.}

\item{rctv_lglDisable}{\verb{reactive logical} Whether to disable the menu.}

\item{rctv_chrDisabledChoices}{\verb{reactive character} Choices to disable.}

\item{rctv_lglOpen}{\verb{reactive logical} Whether to open (\code{TRUE}) or close
(\code{FALSE}) the menu.}
}
\description{
Inline virtual select input server
}
\keyword{internal}
