% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_InlineSelect.R
\name{mod_InlineSelect_UI}
\alias{mod_InlineSelect_UI}
\title{Inline virtual select input}
\usage{
mod_InlineSelect_UI(
  id,
  strLabel = NULL,
  chrChoices = "",
  chrSelected = NULL,
  ...
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strLabel}{\code{character} The initial label for the select input.}

\item{chrChoices}{\code{character} A (named, optionally) vector of choices for
this input.}

\item{chrSelected}{\code{character} The initially selected value (or multiple
values if multiple = TRUE). If not specified then defaults to the first
value for single-select lists and no values for multiple select lists.}

\item{...}{Additional parameters passed on to
\code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}}.}
}
\value{
A \code{\link[htmltools:span]{htmltools::span()}} element containing necessary elements for an
inline \code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}}.
}
\description{
Inline virtual select input
}
\keyword{internal}
