% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable.R
\name{mod_MetricTable_UI}
\alias{mod_MetricTable_UI}
\title{Metric Table Module UI}
\usage{
mod_MetricTable_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and a
\code{\link[gsm.kri:Report_MetricTable]{gsm.kri::Report_MetricTable()}}.
}
\description{
Metric Table Module UI
}
\keyword{internal}
