% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_NavSelect.R
\name{mod_NavSelect_UI}
\alias{mod_NavSelect_UI}
\title{Navigation Select Input UI}
\usage{
mod_NavSelect_UI(id, chrChoices, strLabel = NULL)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrChoices}{\code{character} A (named, optionally) vector of choices for
this input.}

\item{strLabel}{\code{character} The initial label for the select input.}
}
\value{
A \code{\link[bslib:nav_item]{bslib::nav_item()}} with a formatted
\code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}}.
}
\description{
Navigation Select Input UI
}
\keyword{internal}
