% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant_details_ui.R
\name{mod_participant_details_ui}
\alias{mod_participant_details_ui}
\title{Participant Details UI}
\usage{
mod_participant_details_ui(id)
}
\arguments{
\item{id}{The namespace id for this module.}
}
\description{
Participant Details UI
}
