% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_Server.R
\name{mod_ParticipantDetails_Server}
\alias{mod_ParticipantDetails_Server}
\title{Participant Details server}
\usage{
mod_ParticipantDetails_Server(id, fnFetchParticipantData, rctv_strSubjectID)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{fnFetchParticipantData}{\code{function} A function that takes a single
\code{strSubjectID} argument such as "0001", and returns a list with components
\code{metadata} and \code{metric_data}. \code{metadata} should be a named charact vector
(or something that can be coerced to a character vector) of information
about the participant, and \code{metric_data} should be a named list of
data.frames, each of which contains information related to the named
metric.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{SubjectID} of the selected participant.}
}
\description{
Update Participant Details when the selected participant changes.
}
\keyword{internal}
