% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_UI.R
\name{mod_ParticipantDetails_UI}
\alias{mod_ParticipantDetails_UI}
\title{Participant Details UI}
\usage{
mod_ParticipantDetails_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[bslib:layout_columns]{bslib::layout_columns()}} with participant metadata, available
metrics for this participant, and participant domain details.
}
\description{
Participant Details UI
}
\keyword{internal}
