% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDomain_Server.R
\name{mod_ParticipantDomain_Server}
\alias{mod_ParticipantDomain_Server}
\title{Participant Domain Server}
\usage{
mod_ParticipantDomain_Server(id, rctv_lData, rctv_strName)
}
\arguments{
\item{id}{The id for this module in this namespace.}

\item{rctv_lData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns a
named list of dataframes.}

\item{rctv_strName}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the name of an object, such as a particular dataframe in a named
list.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the index(es) of any rows selected by the
user.
}
\description{
Display a specific (named) table, or a placeholder.
}
\keyword{internal}
