% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDomainSummary_Server.R
\name{mod_ParticipantDomainSummary_Server}
\alias{mod_ParticipantDomainSummary_Server}
\title{Participant Domain Summary server}
\usage{
mod_ParticipantDomainSummary_Server(id, rctv_lParticipantDomainData)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A string identifying the selected summary as a \code{\link[shiny:reactive]{shiny::reactive()}}.
}
\description{
Participant Domain Summary server
}
\keyword{internal}
