% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDomainSummary_UI.R
\name{mod_ParticipantDomainSummary_UI}
\alias{mod_ParticipantDomainSummary_UI}
\title{Participant domain summary card}
\usage{
mod_ParticipantDomainSummary_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with either a placeholder, or a clickable list of
available domain data for this participant.
}
\description{
Participant domain summary card
}
\keyword{internal}
