% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantMetricSummary_Server.R
\name{mod_ParticipantMetricSummary_Server}
\alias{mod_ParticipantMetricSummary_Server}
\title{Participant Metric Summary server}
\usage{
mod_ParticipantMetricSummary_Server(id, rctv_lParticipantMetricData)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_lParticipantMetricData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}}
object that returns a list of dataframes for a given participant, one for
each metric that is available.}
}
\value{
A string identifying the selected summary as a \code{\link[shiny:reactive]{shiny::reactive()}}.
}
\description{
Participant Metric Summary server
}
\keyword{internal}
