% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantMetricSummary_UI.R
\name{mod_ParticipantMetricSummary_UI}
\alias{mod_ParticipantMetricSummary_UI}
\title{Participant metric summary card}
\usage{
mod_ParticipantMetricSummary_UI(id)
}
\arguments{
\item{id}{The id for this module in this namespace.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with either a placeholder, or a clickable list of
available metric data for this participant.
}
\description{
Participant metric summary card
}
\keyword{internal}
