% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugin.R
\name{mod_Plugin_Server}
\alias{mod_Plugin_Server}
\title{Plugin Wrapper Server}
\usage{
mod_Plugin_Server(
  lPlugin,
  id,
  lDataModel,
  l_rctvDomains,
  l_rctvDomainHashes,
  l_rctvInputs,
  chrInputNamesPretty
)
}
\arguments{
\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}

\item{id}{\code{character} The id for this element.}

\item{lDataModel}{\code{list} Named list of the standard gsm dataframes
(\code{dfAnalyticsInput}, \code{dfBounds}, \code{dfGroups}, \code{dfMetrics}, and \code{dfResults}).}

\item{l_rctvDomains}{\code{list} A named list of lists of \code{\link[shiny:reactive]{shiny::reactive()}}
objects. Each list contains a "study" element and a "selection" element,
each of which returns a domain dataframe.}

\item{l_rctvDomainHashes}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects,
each of which returns the \code{\link[rlang:hash]{rlang::hash()}} of a domain dataframe.}

\item{l_rctvInputs}{\code{list} A named list of \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} objects,
each of which returns the current value of an input or input-like variable.}

\item{chrInputNamesPretty}{\code{character} A vector of names to display for
missing inputs.}
}
\description{
Plugin Wrapper Server
}
\keyword{internal}
