% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugins_UI.R
\name{mod_Plugins_UI}
\alias{mod_Plugins_UI}
\title{Plugins Wrapper UI}
\usage{
mod_Plugins_UI(id, lPlugins = NULL)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}
}
\value{
Plugins UI elements, either as a tab (if there is only one plugin)
or in a drop-down list (for multiple plugins).
}
\description{
Plugins Wrapper UI
}
\keyword{internal}
