% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PotentialChoice.R
\name{mod_PotentialChoice_Server}
\alias{mod_PotentialChoice_Server}
\title{Potential Choice Select Input Server}
\usage{
mod_PotentialChoice_Server(id, chrChoices, rctv_strSelection)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrChoices}{\code{character} A (named, optionally) vector of choices for
this input.}

\item{rctv_strSelection}{\code{reactiveVal} A \code{reactiveVal} from the parent scope
that this module will update with its selection.}
}
\value{
No returns, but this server creates \code{\link[shiny:observe]{shiny::observe()}} elements to
keep inputs in sync.
}
\description{
Manages the state of a select input that may or may not be rendered in the
UI.
}
\keyword{internal}
