% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PreparePrevalenceData.R
\name{mod_PreparePrevalenceData_Server}
\alias{mod_PreparePrevalenceData_Server}
\title{Reactively prepare data for prevalence plots}
\usage{
mod_PreparePrevalenceData_Server(
  id,
  strLevel,
  l_rctvActive,
  rctv_dfData,
  rctv_strGroupLevel,
  ...
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strLevel}{\code{character} The grouping level for this data (one of
\code{"Study"}, \code{"Group"} or \code{"Participant"}, generally).}

\item{l_rctvActive}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a populated character vector or something that becomes
\code{NULL} when passed through \code{\link[=NullifyEmpty]{NullifyEmpty()}}.}

\item{rctv_dfData}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a generic dataframe.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}
}
\value{
A reactive that returns the prepared \code{data.frame}, or \code{NULL}.
}
\description{
Reactively prepare data for prevalence plots
}
\keyword{internal}
