% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlot.R
\name{mod_PrevalencePlot_Server}
\alias{mod_PrevalencePlot_Server}
\title{Server side of the plot part of the prevalence plot module}
\usage{
mod_PrevalencePlot_Server(id, rctv_dfDomain_Combined, rctv_strCategory)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfDomain_Combined}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}}
object that returns a study-, group-, and/or participant-level domain
dataframe.}

\item{rctv_strCategory}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the currently selected category.}
}
\description{
Server side of the plot part of the prevalence plot module
}
\keyword{internal}
