% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlotCard.R
\name{mod_PrevalencePlotCard_Server}
\alias{mod_PrevalencePlotCard_Server}
\title{Server side of prevalence plot module}
\usage{
mod_PrevalencePlotCard_Server(
  id,
  rctv_dfDomain,
  rctv_dfDomain_Group,
  rctv_dfDomain_Study,
  rctv_strGroupLevel,
  rctv_strGroupID,
  rctv_strSubjectID,
  l_rctvDomainLoaded
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}

\item{rctv_dfDomain_Group}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns a group-level domain dataframe.}

\item{rctv_dfDomain_Study}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns a study-level domain dataframe.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{l_rctvDomainLoaded}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects.
The list contains a "Study" element and an element per group level, each of
which returns a \code{logical} (\code{FALSE} if the data has not been accessed,
\code{TRUE} if it has).}
}
\description{
Server side of prevalence plot module
}
\keyword{internal}
