% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlotCard.R
\name{mod_PrevalencePlotCard_UI}
\alias{mod_PrevalencePlotCard_UI}
\title{Interactive plots of prevalence of categorical variables}
\usage{
mod_PrevalencePlotCard_UI(id, strPlotTitle = NULL)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an interactive prevalence plot.
}
\description{
Interactive plots of prevalence of categorical variables
}
\keyword{internal}
