% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_RAGPillSet.R
\name{mod_RAGPillSet_UI}
\alias{mod_RAGPillSet_UI}
\title{Clickable KRI Count Pills}
\usage{
mod_RAGPillSet_UI(id, intRed = NULL, intAmber = NULL, strLabel = "KRIs")
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{intRed}{\code{integer} The number of groups with at least one red flag.}

\item{intAmber}{\code{integer} The number of groups with at least one amber flag.}

\item{strLabel}{\code{character} The label of a field.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{\link[=mod_TogglePill_UI]{mod_TogglePill_UI()}} items, one for red
flags, and one for amber flags.
}
\description{
Clickable KRI Count Pills
}
\keyword{internal}
