% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlotSet_Server.R
\name{mod_ScatterPlotSet_Server_MetricID}
\alias{mod_ScatterPlotSet_Server_MetricID}
\title{ScatterPlotSet Module Server (MetricID)}
\usage{
mod_ScatterPlotSet_Server_MetricID(id)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the id of the selected metric.
}
\description{
ScatterPlotSet Module Server (MetricID)
}
\keyword{internal}
