% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlotSet.R
\name{mod_ScatterPlotSet_UI}
\alias{mod_ScatterPlotSet_UI}
\title{ScatterPlot Set Module UI}
\usage{
mod_ScatterPlotSet_UI(id, dfMetrics)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created with \code{\link[gsm.reporting:MakeMetric]{gsm.reporting::MakeMetric()}}.}
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \code{\link[bslib:layout_column_wrap]{bslib::layout_column_wrap()}}
of plots, and the output of \code{\link[=out_DetectCardClicks]{out_DetectCardClicks()}} (the JavaScript
necessary to detect that the plot has been clicked).
}
\description{
ScatterPlot Set Module UI
}
\keyword{internal}
