% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectDataCategory.R
\name{mod_SelectDataCategory_Server}
\alias{mod_SelectDataCategory_Server}
\title{Choose categories and update the dropdown}
\usage{
mod_SelectDataCategory_Server(id, rctv_dfDomain)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}
}
\description{
Choose categories and update the dropdown
}
\keyword{internal}
