% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectState.R
\name{mod_SelectState_Server}
\alias{mod_SelectState_Server}
\title{Select State Server}
\usage{
mod_SelectState_Server(
  id,
  rctv_strDependent,
  l_chrChoices,
  lglUpdateLabel,
  rctv_strCurrentSelection
)
}
\arguments{
\item{id}{\code{character} The module ID.}

\item{rctv_strDependent}{\code{reactive} The reactive string value that this
module's selection depends on.}

\item{l_chrChoices}{\code{list} A named list where names correspond to values from
\code{rctv_strDependent} and values are the character vectors of choices.}

\item{lglUpdateLabel}{\code{logical} A flag indicating whether to update the
select input's label with the value of \code{rctv_strDependent}.}

\item{rctv_strCurrentSelection}{\code{reactive} The currently selected value from
the select input UI.}
}
\value{
A \code{list} of \code{reactives}: \code{rctv_chrChoices}, \code{rctv_strLabel}, and
\code{rctv_strSelected}.
}
\description{
A testable module that calculates the state of a cascading select input.
}
\keyword{internal}
