% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_site_details_server.R
\name{mod_site_details_server}
\alias{mod_site_details_server}
\title{Site Details Server}
\usage{
mod_site_details_server(
  id,
  dfMetrics,
  dfGroups,
  dfAnalyticsInput,
  rctv_strSiteID,
  rctv_strMetricID
)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strSiteID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strMetricID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the selected \code{MetricID}.}
}
\description{
Site Details Server
}
