% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_site_details_ui.R
\name{mod_site_details_ui}
\alias{mod_site_details_ui}
\title{Site Details UI}
\usage{
mod_site_details_ui(id)
}
\arguments{
\item{id}{The namespace id for this module.}
}
\description{
Site Details UI
}
