% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteDetails_Server.R
\name{mod_SiteDetails_Server}
\alias{mod_SiteDetails_Server}
\title{Site Details Server}
\usage{
mod_SiteDetails_Server(
  id,
  dfMetrics,
  dfGroups,
  dfAnalyticsInput,
  rctv_strSiteID,
  rctv_strMetricID
)
}
\arguments{
\item{id}{The id for this module in this namespace.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}
}
\description{
Site Details Server
}
\keyword{internal}
