% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteDetails_UI.R
\name{mod_SiteDetails_UI}
\alias{mod_SiteDetails_UI}
\title{Site Details UI}
\usage{
mod_SiteDetails_UI(id)
}
\arguments{
\item{id}{The id for this module or widget in this namespace.}
}
\description{
Site Details UI
}
\keyword{internal}
