% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteParticipants_Server.R
\name{mod_SiteParticipants_Server}
\alias{mod_SiteParticipants_Server}
\title{Site Participants Server}
\usage{
mod_SiteParticipants_Server(
  id,
  rctv_strGroupID,
  rctv_strSubjectID,
  rctv_dfAnalyticsInput,
  rctv_lColumnNames
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} with the id of the participant selected in
the table.
}
\description{
Site Participants Server
}
\keyword{internal}
