% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteParticipants_UI.R
\name{mod_SiteParticipants_UI}
\alias{mod_SiteParticipants_UI}
\title{Site Participants UI}
\usage{
mod_SiteParticipants_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with a placeholder or a clickable table of site
participants.
}
\description{
Site Participants UI
}
\keyword{internal}
