% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_StudyInformation.R
\name{mod_StudyInformation_UI}
\alias{mod_StudyInformation_UI}
\title{Study Information Card UI}
\usage{
mod_StudyInformation_UI(id, dfGroups, dfResults)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with overall study metadata.
}
\description{
Study Information Card UI
}
\keyword{internal}
