% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TimeSeries_UI.R
\name{mod_TimeSeries_UI}
\alias{mod_TimeSeries_UI}
\title{TimeSeries Module UI}
\usage{
mod_TimeSeries_UI(id, strPlotTitle = NULL)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and a
\code{\link[=Widget_TimeSeriesOutput]{Widget_TimeSeriesOutput()}}.
}
\description{
TimeSeries Module UI
}
\keyword{internal}
