% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TogglePill.R
\name{mod_TogglePill_UI}
\alias{mod_TogglePill_UI}
\title{Toggle Pill Module UI}
\usage{
mod_TogglePill_UI(id, strLabel, strColorCode)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strLabel}{\code{character} The label of a field.}

\item{strColorCode}{\code{character} The hex code (such as \code{"#FFFFFF"}) for a
color.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} with a specialized \code{\link[shiny:actionLink]{shiny::actionLink()}}
and the necessary CSS and JavaScript to manage the link.
}
\description{
Toggle Pill Module UI
}
\keyword{internal}
