% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_WidgetPlot.R
\name{mod_WidgetPlot_Server}
\alias{mod_WidgetPlot_Server}
\title{Standardized Plot Widgets as Shiny Modules}
\usage{
mod_WidgetPlot_Server(
  id,
  fn_Widget,
  fn_WidgetOutput,
  rctv_dfResults,
  rctv_lMetric,
  dfGroups,
  rctv_dfBounds,
  rctv_strGroupID,
  ...
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{fn_Widget}{A function that uses \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}} to create
a widget.}

\item{fn_WidgetOutput}{A function that uses
\code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyWidgetOutput()}} to create the shiny version of a widget.}

\item{rctv_dfResults}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a stacked summary of analysis pipeline output.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{rctv_dfBounds}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a set of predicted percentages/rates and upper- and lower-bounds
across the full range of sample sizes/total exposure values for reporting.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{...}{Additional parameters passed to \code{fn_Widget}.}
}
\description{
Standardized Plot Widgets as Shiny Modules
}
\keyword{internal}
