% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_WidgetPlot.R
\name{mod_WidgetPlot_UI}
\alias{mod_WidgetPlot_UI}
\title{Widget Plot Module UI}
\usage{
mod_WidgetPlot_UI(
  id,
  fnHtmlDependency,
  fnWidgetOutput,
  strPlotTitle = NULL,
  strGroupLevel = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{fnHtmlDependency}{\code{function} A function that returns an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} or multiple wrapped in \code{\link[shiny:tagList]{shiny::tagList()}}.}

\item{fnWidgetOutput}{\code{function} An \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}
function.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: \code{NULL}
and taken from \code{dfMetrics$GroupLevel} if available.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and the corresponding
\code{fnWidgetOutput} and dependencies to load it.
}
\description{
Widget Plot Module UI
}
\keyword{internal}
