% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{null_for_none}
\alias{null_for_none}
\title{Replace empty values with NULL}
\usage{
null_for_none(strValue)
}
\arguments{
\item{strValue}{\code{character} The value of a field.}
}
\value{
The value, or \code{NULL} if the value is "None", "All", or an empty
string.
}
\description{
Replace empty values with NULL
}
\keyword{internal}
