% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NullifyEmpty}
\alias{NullifyEmpty}
\title{Replace empty values with NULL}
\usage{
NullifyEmpty(strValue)
}
\arguments{
\item{strValue}{\code{character} The value of a field.}
}
\value{
The value, or \code{NULL} if the value is "None", "All", or an empty
string.
}
\description{
Replace empty values with NULL
}
\keyword{internal}
