% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_method.R
\name{observe_method}
\alias{observe_method}
\title{Update numeric inputs given change to method input.}
\usage{
observe_method(input, session)
}
\arguments{
\item{input}{\code{list} Shiny input}

\item{session}{\code{list} Shiny session}
}
\description{
Update numeric inputs given change to method input.
}
