% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_metric_select.R
\name{observe_metric_select}
\alias{observe_metric_select}
\title{Observe Metric Select}
\usage{
observe_metric_select(snapshot, metric)
}
\arguments{
\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}

\item{metric}{The reactive value provided by the metric input from \code{server}}
}
\description{
Observe Metric Select
}
