% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-length-0-default}
\alias{op-length-0-default}
\alias{\%|0|\%}
\title{Default value for zero length}
\usage{
x \%|0|\% y
}
\arguments{
\item{x}{A value to check for length.}

\item{y}{A default value.}
}
\value{
If \code{x} has a length, \code{x}; otherwise \code{y}.
}
\description{
Utility to easily replace length-0 values with a default value. Similar to
the infix operator (\verb{\%||\%}) from \code{rlang}, which was recently added to base R.
}
\keyword{internal}
