% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_DetectCardClicks.R
\name{out_DetectCardClicks}
\alias{out_DetectCardClicks}
\title{Card click detection JavaScript}
\usage{
out_DetectCardClicks(strContainerID, strInputID)
}
\arguments{
\item{strContainerID}{\code{character} The (namespaced) ID of the target container
(usually a div).}

\item{strInputID}{\code{character} An ID to use for the Shiny input created by
this module or used by this JavaScript.}
}
\value{
A \code{\link[shiny:tagList]{shiny::tagList()}} with the JavaScript required to enable card
click detection within the specified container. Clicks will be reported to
the Shiny input named \code{strInputID}.
}
\description{
Card click detection JavaScript
}
\keyword{internal}
