% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_gt.R
\name{out_gtInteractive}
\alias{out_gtInteractive}
\title{Interactive gt Table}
\usage{
out_gtInteractive(gt_object, selection_mode = "single")
}
\arguments{
\item{gt_object}{A \code{\link[gt:gt]{gt::gt()}} object to use interactively.}

\item{selection_mode}{\emph{Allow row selection}

\verb{scalar<character>} // \emph{default:} \code{NULL}

The \code{selection_mode} options allows users to select rows by clicking them.
When this option is \code{"single"}, clicking another value toggles selection
of the previously selected row off. When this option is \code{"multiple"},
multiple rows can be selected at once. Selected values are available in
Shiny apps when \code{selection_mode} is not \code{NULL} and the table is used in
\code{\link[gt:render_gt]{render_gt()}}.}
}
\value{
A \code{\link[gt:gt]{gt::gt()}} with our standard options.
}
\description{
Interactive gt Table
}
\keyword{internal}
