% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_gt.R
\name{out_gtPlaceholder}
\alias{out_gtPlaceholder}
\title{Placeholder gt Table}
\usage{
out_gtPlaceholder(..., id = NULL)
}
\arguments{
\item{...}{\code{character} Text to describe what is required, as a string,
character vector, or multiple strings or character vectors.}

\item{id}{\code{character} An optional ID for the table.}
}
\value{
A \code{\link[gt:gt]{gt::gt()}} with a placeholder.
}
\description{
Placeholder gt Table
}
\keyword{internal}
