% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_gt.R
\name{out_gtSmartFmtNumbers}
\alias{out_gtSmartFmtNumbers}
\title{Intelligently format numeric data}
\usage{
out_gtSmartFmtNumbers(gt_object, intMaxDecimals = 5L)
}
\arguments{
\item{gt_object}{A \code{\link[gt:gt]{gt::gt()}} object to use interactively.}
}
\value{
A \code{\link[gt:gt]{gt::gt()}} with numeric formatting applied.
}
\description{
Intelligently format numeric data
}
\keyword{internal}
