% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Sidebar.R
\name{out_Inputs}
\alias{out_Inputs}
\title{Card with inputs}
\usage{
out_Inputs(chrDomains, chrMetrics)
}
\arguments{
\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with app inputs.
}
\description{
Card with inputs
}
\keyword{internal}
