% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_KRIColor.R
\name{out_KRIColor}
\alias{out_KRIColor}
\title{KRI Color Div}
\usage{
out_KRIColor(dfResults)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}
}
\value{
A \code{\link[htmltools:builder]{htmltools::div()}} with Red and Amber flag counts.
}
\description{
KRI Color Div
}
\keyword{internal}
