% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_KRIColor.R
\name{out_KRIColor_pill}
\alias{out_KRIColor_pill}
\title{KRI color pill}
\usage{
out_KRIColor_pill(intKRIColorCount, strColor)
}
\arguments{
\item{intKRIColorCount}{\code{integer} A named vector of counts by color.}

\item{strColor}{\code{character} The target color to report about.}
}
\value{
A \code{\link[htmltools:builder]{htmltools::span()}} with the count of the given color, with a class
to match the color.
}
\description{
KRI color pill
}
\keyword{internal}
