% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_MainTabs.R
\name{out_MainTabs}
\alias{out_MainTabs}
\title{Main Tabbed Content}
\usage{
out_MainTabs(chrDomains, dfGroups, dfMetrics, dfResults, lPlugins = NULL)
}
\arguments{
\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created with \code{\link[gsm.reporting:MakeMetric]{gsm.reporting::MakeMetric()}}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created with \code{\link[gsm.reporting:BindResults]{gsm.reporting::BindResults()}}.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}
}
\value{
A list of \code{\link[bslib:nav_panel]{bslib::nav_panel()}} elements, containing the main tabbed
content.
}
\description{
Main Tabbed Content
}
\keyword{internal}
