% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_MetadataList.R
\name{out_MetadataItem}
\alias{out_MetadataItem}
\title{Metadata list item}
\usage{
out_MetadataItem(strLabel, strValue)
}
\arguments{
\item{strLabel}{\code{character} The label of a field.}

\item{strValue}{\code{character} The value of a field.}
}
\value{
A \code{\link[htmltools:div]{htmltools::div()}} with the label, a spacer, and the value.
}
\description{
Metadata list item
}
\keyword{internal}
