% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_MetadataList.R
\name{out_MetadataList}
\alias{out_MetadataList}
\title{Metadata list card}
\usage{
out_MetadataList(chrValues, chrLabels = names(chrValues))
}
\arguments{
\item{chrValues}{\code{character} A vector of values to associate with a vector of
labels.}

\item{chrLabels}{\code{character} A vector of labels for display to the user.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} of label-value pairs, including the CSS
needed to lay them out properly.
}
\description{
Display label-value pairs as a vertical list, with a consistent spacer
between the label and value.
}
\examples{
out_MetadataList(
  c("value 1", "value 2", "value 3"),
  c("Thing 1", "Thing 2", "Thing 3")
)
}
