% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_server.R
\name{out_ParticipantCard_Wrapper}
\alias{out_ParticipantCard_Wrapper}
\title{Participant card}
\usage{
out_ParticipantCard_Wrapper(strCardTitle, divParticipantCore)
}
\arguments{
\item{strCardTitle}{\code{character} A title for the card.}

\item{divParticipantCore}{The core participant content, such as a placeholder
or a collection of rows.}
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} of participant data.
}
\description{
Participant card
}
\keyword{internal}
