% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantMetadata.R
\name{out_ParticipantMetadata_Item}
\alias{out_ParticipantMetadata_Item}
\title{A single-row div or participant data}
\usage{
out_ParticipantMetadata_Item(strValue, strName)
}
\arguments{
\item{strValue}{\code{character} The value of a field.}

\item{strName}{\code{character} The name of a field.}
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} laying out a single row of data.
}
\description{
A single-row div or participant data
}
\keyword{internal}
