% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantMetricSummary.R
\name{out_ParticipantMetricSummary}
\alias{out_ParticipantMetricSummary}
\title{Participant metadata card}
\usage{
out_ParticipantMetricSummary(id, lParticipantMetadata)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{lParticipantMetadata}{\code{list} Named list of data describing a single
participant.}
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} with participant metadata.
}
\description{
Participant metadata card
}
\keyword{internal}
