% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantMetricSummary.R
\name{out_ParticipantMetricSummary_Item}
\alias{out_ParticipantMetricSummary_Item}
\title{A single-row div or participant data}
\usage{
out_ParticipantMetricSummary_Item(id, lParticipantMetadata)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{lParticipantMetadata}{\code{list} Named list of data describing a single
participant.}
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} laying out a single row of data.
}
\description{
A single-row div or participant data
}
\keyword{internal}
