% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantsBySite.R
\name{out_ParticipantsBySite}
\alias{out_ParticipantsBySite}
\title{Table of Participants Based on Site Selection}
\usage{
out_ParticipantsBySite(dfAnalyticsInput, chrColumnNames)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}
}
\description{
Table of Participants Based on Site Selection
}
\keyword{internal}
