% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_SiteParticipants.R
\name{out_SiteParticipants}
\alias{out_SiteParticipants}
\title{Table of Participants Based on Site Selection}
\usage{
out_SiteParticipants(dfAnalyticsInput, chrColumnNames)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}
}
\description{
Table of Participants Based on Site Selection
}
\keyword{internal}
