% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_StudyInformation.R
\name{out_StudyInformation}
\alias{out_StudyInformation}
\title{Study Information Card}
\usage{
out_StudyInformation(lStudy)
}
\arguments{
\item{lStudy}{\code{list} Named list of data describing the overall study.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with overall study metadata.
}
\description{
Study Information Card
}
\keyword{internal}
