% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview_table.R
\name{overview_table_server}
\alias{overview_table_server}
\title{Define a Shiny server function given an assessment}
\usage{
overview_table_server(input, output, session, params)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny server function
}
\description{
Define a Shiny server function given an assessment
}
