% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview_table.R
\name{overview_table_ui}
\alias{overview_table_ui}
\title{Define a Shiny UI function given an assessment}
\usage{
overview_table_ui(id)
}
\arguments{
\item{assessment}{\code{list} assessment specification}
}
\value{
\code{function} Shiny UI function
}
\description{
Define a Shiny UI function given an assessment
}
