% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant_details_ui.R
\name{participant_details_ui}
\alias{participant_details_ui}
\title{Participant Details UI}
\usage{
participant_details_ui(id)
}
\arguments{
\item{id}{The namespace id}
}
\description{
Participant Details UI
}
