% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_metric_summary_tag_list.R
\name{participant_metric_summary_tag_list}
\alias{participant_metric_summary_tag_list}
\title{Creates taglist to summarize participants at a particular site}
\usage{
participant_metric_summary_tag_list(id, participant, snapshot)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{participant}{A combined data set with the required metrics.}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{'Enrollment',}
\keyword{'Queries'}
\keyword{=}
\keyword{dfENROLL'}
\keyword{dfQUERY'}
\keyword{internal}
