% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_status_nest_list.R
\name{participant_status_nest_list}
\alias{participant_status_nest_list}
\title{Creates taglist to summarize participants at a particular site}
\usage{
participant_status_nest_list(participant_list = NULL)
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{internal}
