% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_ParticipantStatusNestedList.R
\name{participantStudyNestedList}
\alias{participantStudyNestedList}
\title{Creates taglist to summarize participants at a particular site}
\usage{
participantStudyNestedList(participant_list = NULL)
}
\arguments{
\item{combinedData}{\code{data.frame} A combined data set with the required metrics.}
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{internal}
