% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{pkgloadHack}
\alias{pkgloadHack}
\title{pkgload as a dependency}
\usage{
pkgloadHack()
}
\description{
This is only a placeholder function that has no actual use outside of existing to remove the \code{devtools::check()} flag that a dependency is not used.
We need this as a dependency to host the app on shinyapps.io - this function tricks R into thinking it's a dependency that is being used.
}
\keyword{internal}
