% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlot.R
\name{PlotPrevalencePlot}
\alias{PlotPrevalencePlot}
\title{Generate the prevalence ggplot}
\usage{
PlotPrevalencePlot(dfPrevalence, strCategory)
}
\arguments{
\item{dfPrevalence}{\code{data.frame} The prepared data.}

\item{strCategory}{\code{character} A category to focus on.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Generate the prevalence ggplot
}
\keyword{internal}
