% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_GetDependencySources}
\alias{plugin_GetDependencySources}
\title{Get Plugin Package Dependency Sources}
\usage{
plugin_GetDependencySources(lPluginDefinition)
}
\arguments{
\item{lPluginDefinition}{\code{list} A named list with required elements
\code{strTitle}, \code{fnUI}, and \code{fnServer}, and optional fields \code{packages} and
\code{lConfig}. Usually generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
A character vector of package sources, such as "ggplot2" (to install
from CRAN) or
"url::https://safetygraphics.r-universe.dev/src/contrib/safetyCharts_0.4.0.tar.gz"
(to install from a specific URL on r-universe).
}
\description{
Retrieve a vector of sources for plugin package dependencies, to make it
easier to install those sources. This function is intended for use in
automated deployment systems, such as GitHub Actions.
}
