% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_InstallDependencySources}
\alias{plugin_InstallDependencySources}
\title{Install Plugin Package Dependencies}
\usage{
plugin_InstallDependencySources(lPluginDefinition)
}
\arguments{
\item{lPluginDefinition}{\code{list} A named list with required elements
\code{strTitle}, \code{fnUI}, and \code{fnServer}, and optional fields \code{packages} and
\code{lConfig}. Usually generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
\code{lPluginDefinition}, invisibly. This function is called for its side
effects.
}
\description{
Plugins can have additional dependencies that are invisible to the usual
CI/CD pipelines. Use this function to install any such dependencies using the
pak package.
}
