% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_LoadDependencies}
\alias{plugin_LoadDependencies}
\title{Load Plugin Dependencies}
\usage{
plugin_LoadDependencies(lPlugin)
}
\arguments{
\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
\code{lPlugin}, invisibly. This function is called for its side effects.
}
\description{
Load the package dependencies of a plugin. This is designed to be used in an
\code{app.R} file to ensure that the dependencies are detected by packages like
rsconnect.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
plugin_LoadDependencies(list(packages = list(list(name = "gsm.app"))))
\dontshow{\}) # examplesIf}
}
