% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_Prepare}
\alias{plugin_Prepare}
\title{Finish Plugin Definition}
\usage{
plugin_Prepare(
  lPlugin,
  lWorkflows = list(),
  ...,
  envCall = rlang::caller_env()
)
}
\arguments{
\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}

\item{lWorkflows}{\code{list} An optional named list of workflows to run in order
to translate app domain data into the format required by the plugin (as
defined in the plugin \code{spec}). These workflows are ran before the plugin's
\code{spec} is applied, and are meant to serve as a bridge between your domain
data and the plugin's expected domains.}

\item{...}{Additional named arguments passed on to \code{fnShinyUI} and/or
\code{fnShinyServer} by name.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin}.
}
\description{
Finish Plugin Definition
}
\keyword{internal}
