% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_Read}
\alias{plugin_Read}
\title{Read a Plugin Definition}
\usage{
plugin_Read(strPath)
}
\arguments{
\item{strPath}{The directory that contains the plugin.}
}
\value{
A \code{list} with elements \code{strTitle}, \code{fnUi}, \code{fnServer}, and
(optionally) \code{lConfig}, read from the YAML file in \code{strPath}. As a side
effect, any \code{R} files in \code{strPath} are also loaded using \code{source()}.
}
\description{
Plugins are defined by a named \code{list} with elements \code{strTitle} (a title to
show in the plugin menu), \code{fnUI} (a function or the name of a function to use
as the plugin's module UI), and \code{fnServer} (a function or the name of a
function to use as the plugin's module server). Use \code{plugin_Read()} to read
these element definitions and any R files in the same directory.
}
\examples{
aePlugin <- plugin_Read(system.file("plugins", "AE", package = "gsm.app"))
aePlugin
}
