% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_Read_yaml_has_all}
\alias{plugin_Read_yaml_has_all}
\title{Check that a plugin definition has all required fields}
\usage{
plugin_Read_yaml_has_all(
  lPluginDefinition,
  chrRequiredFields = c("strTitle", "fnUI", "fnServer"),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{lPluginDefinition}{\code{list} A named list with required elements
\code{strTitle}, \code{fnUI}, and \code{fnServer}, and optional field \code{lConfig}.}

\item{chrRequiredFields}{\code{character} A vector of expected fields in a Plugin
definition.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
A list with a plugin definition.
}
\description{
Check that a plugin definition has all required fields
}
\keyword{internal}
