% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_ReadYaml}
\alias{plugin_ReadYaml}
\title{Process a Plugin YAML}
\usage{
plugin_ReadYaml(chrPluginFiles, envCall = rlang::caller_env())
}
\arguments{
\item{chrPluginFiles}{\code{character} The files in a plugin directory.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
A list with the validated plugin definition.
}
\description{
Process a Plugin YAML
}
\keyword{internal}
