% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_ValidateDefinition}
\alias{plugin_ValidateDefinition}
\title{Validate Plugin Definition}
\usage{
plugin_ValidateDefinition(lPlugin, envCall = rlang::caller_env())
}
\arguments{
\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin}.
}
\description{
Validate Plugin Definition
}
\keyword{internal}
