% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_ValidateMeta}
\alias{plugin_ValidateMeta}
\title{Validate Plugin Definition Meta}
\usage{
plugin_ValidateMeta(lPluginMeta, envCall = rlang::caller_env())
}
\arguments{
\item{lPluginMeta}{\code{list} The plugin metadata.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin$meta}.
}
\description{
Validate Plugin Definition Meta
}
\keyword{internal}
