% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_ValidatePackages}
\alias{plugin_ValidatePackages}
\title{Validate Plugin Definition Packages}
\usage{
plugin_ValidatePackages(lPluginPackages, envCall = rlang::caller_env())
}
\arguments{
\item{lPluginPackages}{\code{list} The names and sources of packages used by the
plugin.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin$packages}.
}
\description{
Validate Plugin Definition Packages
}
\keyword{internal}
