% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_ValidateReqs}
\alias{plugin_ValidateReqs}
\title{Validate Plugin Definition Required Inputs}
\usage{
plugin_ValidateReqs(chrRequiredInputs, envCall = rlang::caller_env())
}
\arguments{
\item{chrRequiredInputs}{\code{character} An optional vector of any inputs
("metric", "site", "participant", or "domain") that must have a non-empty
value before the plugin can load. "None" and "All" count as "empty" for
this check. If the user has not set a value for that input, the app will
display a placeholder instructing the user to make a selection.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin$required_inputs}.
}
\description{
Validate Plugin Definition Required Inputs
}
\keyword{internal}
