% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Define.R
\name{plugin_ValidateShiny}
\alias{plugin_ValidateShiny}
\title{Validate Plugin Definition Shiny}
\usage{
plugin_ValidateShiny(lPluginShiny, envCall = rlang::caller_env())
}
\arguments{
\item{lPluginShiny}{\code{list} The UI and Server function of the plugin.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated \code{lPlugin$shiny}.
}
\description{
Validate Plugin Definition Shiny
}
\keyword{internal}
