% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlot.R
\name{PreparePrevalenceCounts}
\alias{PreparePrevalenceCounts}
\title{Prepare the final prevalence data for plotting}
\usage{
PreparePrevalenceCounts(dfDomain_Combined, strCategory, chrTopValues)
}
\arguments{
\item{dfDomain_Combined}{\code{data.frame} Values of \code{strCategory} for each
\code{VizLevel}.}

\item{chrTopValues}{\code{character} The values to show (other values will be
wrapped into "Other").}
}
\value{
A ready-to-plot data.frame.
}
\description{
Prepare the final prevalence data for plotting
}
\keyword{internal}
