% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PreparePrevalenceData.R
\name{PreparePrevalenceData}
\alias{PreparePrevalenceData}
\title{Prepare data for prevalence plots}
\usage{
PreparePrevalenceData(df, strGroupLevel, strLevel)
}
\arguments{
\item{df}{\code{data.frame} A data frame to wrangle.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: \code{NULL}
and taken from \code{dfMetrics$GroupLevel} if available.}

\item{strLevel}{\code{character} The grouping level for this data (one of
\code{"Study"}, \code{"Group"} or \code{"Participant"}, generally).}
}
\value{
The prepared \code{data.frame}, or \code{NULL}.
}
\description{
Prepare data for prevalence plots
}
\keyword{internal}
