% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PullTopValues.R
\name{PullTopValues}
\alias{PullTopValues}
\title{Find frequent values}
\usage{
PullTopValues(dfDomain, strCategory)
}
\arguments{
\item{dfDomain}{\code{data.frame} Detailed data about a specific domain, such as
"Adverse Events".}

\item{strCategory}{\code{character} A category to focus on.}
}
\value{
An ordered character vector of the top 6 categories (or fewer if
there are fewer than 6 total categories in the data).
}
\description{
Find frequent values
}
\keyword{internal}
