% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snapshot.R
\name{read_snapshot}
\alias{read_snapshot}
\title{Read and assemble snapshot from \code{inst}.}
\usage{
read_snapshot()
}
\value{
\code{list} gsm snapshot object.
}
\description{
Read and assemble snapshot from \code{inst}.
}
\examples{
\dontrun{
snapshot <- read_snapshot()
}
}
