% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\name{RecalculateDomainData}
\alias{RecalculateDomainData}
\title{Update Date-based Fields}
\usage{
RecalculateDomainData(df, strDomainID, dSnapshotDate = NULL)
}
\arguments{
\item{df}{\code{data.frame} A data frame to wrangle.}

\item{strDomainID}{\code{character} The domain data.frame to load.}

\item{dSnapshotDate}{\code{Date} The date of a data snapshot.}
}
\value{
The updated data.frame.
}
\description{
This function allows me to store domain data from a single snapshot and then
back calculate things that are cumulative within snapshots.
}
\keyword{internal}
