% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sample_app.R
\name{run_sample_app}
\alias{run_sample_app}
\title{Run the App with Sample Data}
\usage{
run_sample_app()
}
\value{
An object that represents the app. Printing the object or passing it
to \code{\link[shiny:runApp]{shiny::runApp()}} will run the app.
}
\description{
Launch a version of the Shiny app built by this package, using sample data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_sample_app()
\dontshow{\}) # examplesIf}
}
