% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\docType{data}
\name{sample_dfAnalyticsInput}
\alias{sample_dfAnalyticsInput}
\title{Analytics Input Dataset}
\format{
A tibble with 6490 rows and
8 columns:
\describe{
\item{MetricID}{metric identifier}
\item{SubjectID}{unique subject identifier}
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{numerator for this individual}
\item{Denominator}{denominator for this individual}
\item{Metric}{calculated rate/metric value}
\item{SnapshotDate}{The snapshot with which this data is associated. We currently only use the most recent SnapshotDate}
}
}
\source{
Generated from data in the \code{clindata} package, using the
\code{gsm.mapping} package.
}
\usage{
sample_dfAnalyticsInput
}
\description{
Summary metric data for the "Group Subjects" module on the "Metric Details"
tab.
}
\seealso{
Other sample data: 
\code{\link{sample_dfBounds}},
\code{\link{sample_dfGroups}},
\code{\link{sample_dfMetrics}},
\code{\link{sample_dfResults}},
\code{\link{sample_fnCountData}()},
\code{\link{sample_fnFetchData}()}
}
\concept{sample data}
\keyword{datasets}
