% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\docType{data}
\name{sample_dfResults}
\alias{sample_dfResults}
\title{KRI Results Dataset}
\format{
A tibble with 1073 rows and
10 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{number of flags in group}
\item{Denominator}{total subjects in group}
\item{Metric}{calculated rate/metric value}
\item{Score}{statistical score}
\item{Flag}{ordinal flag to be applied}
\item{MetricID}{unique metric identifier}
\item{StudyID}{unique study identifier}
\item{SnapshotDate}{date of the snapshot}
}
}
\source{
Generated from data in the \code{clindata} package, using the
\code{gsm.mapping} package.
}
\usage{
sample_dfResults
}
\description{
Information about the status of each group in the sample data.
}
\seealso{
Other sample data: 
\code{\link{sample_dfAnalyticsInput}},
\code{\link{sample_dfBounds}},
\code{\link{sample_dfGroups}},
\code{\link{sample_dfMetrics}},
\code{\link{sample_fnCountData}()},
\code{\link{sample_fnFetchData}()}
}
\concept{sample data}
\keyword{datasets}
