% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Define Server}
\usage{
server(input, output, session, snapshot)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}
}
\description{
Define Server
}
