% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugin.R
\name{SimplifyArgs}
\alias{SimplifyArgs}
\title{Simplify Function Arguments for Comparison}
\usage{
SimplifyArgs(chrArgs)
}
\arguments{
\item{chrArgs}{\code{character} A vector of argument names to use in comparisons.}
}
\value{
A \code{character} vector of lowercase argument names without prefixes or
special characters.
}
\description{
Simplify Function Arguments for Comparison
}
\keyword{internal}
