% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_details_metric_meta_data_list.R
\name{site_details_metric_meta_data_list}
\alias{site_details_metric_meta_data_list}
\title{Creates taglist to summarize participants at a particular site}
\usage{
site_details_metric_meta_data_list(config_params = NULL, meta_workflow = NULL)
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{internal}
