% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_site_details_server.R
\name{site_details_server}
\alias{site_details_server}
\title{Site Details Server}
\usage{
site_details_server(id, snapshot, site, metric)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}

\item{site}{The reactive value provided by the site input from \code{server}}

\item{metric}{The reactive value provided by the metric input from \code{server}}
}
\description{
Site Details Server
}
