% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{splitByGrouping}
\alias{splitByGrouping}
\title{Split a data frame by a grouping variable.}
\usage{
splitByGrouping(df, strGroupingVar, strValueVar)
}
\arguments{
\item{df}{\code{data.frame} A data frame to wrangle.}

\item{strGroupingVar}{\code{character} The name of the grouping variable.}

\item{strValueVar}{\code{character} The name of the value variable.}
}
\value{
A named list of values.
}
\description{
Split a data frame by a grouping variable.
}
\keyword{internal}
