% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{srvr_CombineHashes}
\alias{srvr_CombineHashes}
\title{Merge Used Hashes}
\usage{
srvr_CombineHashes(l_rctvDomainHashes, lSpec)
}
\arguments{
\item{l_rctvDomainHashes}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects,
each of which returns the \code{\link[rlang:hash]{rlang::hash()}} of a domain dataframe.}

\item{lSpec}{\code{list} A named list defining the data domains required by the
plugin, where the names are the names of the domains and the elements are
column definitions.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} that returns a string with the hashes of the
loaded domains concatenated with "-".
}
\description{
Merge Used Hashes
}
\keyword{internal}
