% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{srvr_CompileUnsetInputs}
\alias{srvr_CompileUnsetInputs}
\title{Figure out which required inputs aren't available}
\usage{
srvr_CompileUnsetInputs(chrRequiredInputs, l_rctvInputs, chrInputNamesPretty)
}
\arguments{
\item{chrRequiredInputs}{\code{character} A vector of inputs required by a Plugin.}

\item{l_rctvInputs}{\code{list} A named list of \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} objects,
each of which returns the current value of an input or input-like variable.}

\item{chrInputNamesPretty}{\code{character} A vector of names to display for
missing inputs.}
}
\value{
A character vector of missing inputs.
}
\description{
Figure out which required inputs aren't available
}
\keyword{internal}
