% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dfDomains.R
\name{srvr_DomainHash}
\alias{srvr_DomainHash}
\title{Reactive domain df}
\usage{
srvr_DomainHash(
  rctvDomain,
  strDomainID,
  rctv_dSnapshotDate,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID
)
}
\arguments{
\item{strDomainID}{\code{character} The domain data.frame to load.}

\item{rctv_dSnapshotDate}{\verb{reactive Date} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the date of a data snapshot.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} that generates a hash of a domain df.
}
\description{
Reactive domain df
}
\keyword{internal}
