% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_l_rctvDomains}
\alias{srvr_l_rctvDomains}
\title{List of reactive domain dfs}
\usage{
srvr_l_rctvDomains(fnFetchData, chrDomains, rctv_strSiteID, rctv_strSubjectID)
}
\arguments{
\item{fnFetchData}{\code{function} A function that takes a \code{strDomain} argument
and optional \code{strSiteID} and \code{strSubjectID} and returns a data.frame. See
\code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The returned data.frame contains
information about the named domain.}

\item{chrDomains}{\code{character} A vector of names of domain data.frames to show
in the app. Supported domains are shown in the default value.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A list of \code{\link[shiny:reactive]{shiny::reactive()}}s with domain dfs.
}
\description{
List of reactive domain dfs
}
\keyword{internal}
