% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_MetricDetails.R
\name{srvr_MetricDetails}
\alias{srvr_MetricDetails}
\title{Temporary Metric Details wrapper}
\usage{
srvr_MetricDetails(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  rctv_strMetricID,
  rctv_strPrimaryNavBar,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  input,
  output,
  session
)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data. This
data.frame is created by binding together the \code{lAnalysis} "Analysis_Input"
list of data.frames output of analysis workflows, with a \code{MetricID} column
for the name of each list.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created with \code{\link[gsm.reporting:MakeBounds]{gsm.reporting::MakeBounds()}}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting. Created with \code{\link[gsm.reporting:MakeMetric]{gsm.reporting::MakeMetric()}}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created with \code{\link[gsm.reporting:BindResults]{gsm.reporting::BindResults()}}.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} for \code{rctv_strGroupDetailsParticipant} from
\code{\link[=mod_GroupDetails_Server]{mod_GroupDetails_Server()}}.
}
\description{
This really should be a full module, but this lets me "hide" some server code
so it isn't all inside \code{\link[=gsmApp_Server]{gsmApp_Server()}}.
}
\keyword{internal}
