% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PopulateParticipantSelect.R
\name{srvr_PopulateParticipantSelect}
\alias{srvr_PopulateParticipantSelect}
\title{Initialize Participant Select}
\usage{
srvr_PopulateParticipantSelect(
  chrParticipantIDs,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\description{
Initialize Participant Select
}
\keyword{internal}
