% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PullTopValues.R
\name{srvr_PullTopValues}
\alias{srvr_PullTopValues}
\title{Find frequent values reactively}
\usage{
srvr_PullTopValues(rctv_dfDomain, rctv_strCategory)
}
\arguments{
\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}

\item{rctv_strCategory}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the currently selected category.}
}
\value{
A reactive that returns an ordered character vector of the top 6
categories (or fewer if there are fewer than 6 total categories in the
data).
}
\description{
Find frequent values reactively
}
\keyword{internal}
