% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_rctv_chrParticipantIDs}
\alias{srvr_rctv_chrParticipantIDs}
\title{Reactive participant IDs}
\usage{
srvr_rctv_chrParticipantIDs(dfAnalyticsInput, rctv_strGroupID)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} of dfMetrics as a list, filtered by metric.
}
\description{
Reactive participant IDs
}
\keyword{internal}
