% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_Reset}
\alias{srvr_Reset}
\title{Reset app}
\usage{
srvr_Reset(
  dfMetrics,
  dfParticipantGroups,
  rctv_intReset,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfParticipantGroups}{\code{data.frame} Unique \code{SubjectID} and \code{GroupID}
combos from \code{dfAnalyticsInput}.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that resets the app when the reset input
changes.
}
\description{
Reset app
}
\keyword{internal}
