% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{srvr_RunReactiveWorkflows}
\alias{srvr_RunReactiveWorkflows}
\title{Run Workflows Reactively}
\usage{
srvr_RunReactiveWorkflows(
  l_rctvDomains,
  lWorkflows,
  l_rctvDomainHashes,
  rctv_bPluginReady
)
}
\arguments{
\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}

\item{l_rctvDomainHashes}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects,
each of which returns the \code{\link[rlang:hash]{rlang::hash()}} of a domain dataframe.}
}
\value{
A named list of the final output from \code{lWorkflows}, where the names
are the names of the final results.
}
\description{
Run Workflows Reactively
}
\keyword{internal}
