% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dfDomains.R
\name{srvr_ShowConditionMessage}
\alias{srvr_ShowConditionMessage}
\title{Modal dialog for errors etc}
\usage{
srvr_ShowConditionMessage(
  cnd,
  strTitle,
  chrPreMessage = character(),
  chrPostMessage = character(),
  strDismiss = "Dismiss"
)
}
\arguments{
\item{cnd}{\code{condition} The condition object.}

\item{strTitle}{\verb{length-1 character} A title for the modal popup.}

\item{chrPreMessage}{\code{character} One more more messages to show before the
condition message.}

\item{chrPostMessage}{\code{character} One or more messages to show after the
condition message.}

\item{strDismiss}{\verb{length-1 character} Text to show on the dismiss button.}
}
\value{
A \code{\link[shiny:modalDialog]{shiny::modalDialog()}} with the designated text, launched via
\code{\link[shiny:showModal]{shiny::showModal()}}.
}
\description{
Modal dialog for errors etc
}
\keyword{internal}
