% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_SyncSelectInput}
\alias{srvr_SyncSelectInput}
\title{Synchronize selectInput with reactive}
\usage{
srvr_SyncSelectInput(id, rctv_strValue, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strValue}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value of a field.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that triggers \code{\link[shiny:updateSelectInput]{shiny::updateSelectInput()}} when
the connected reactive changes.
}
\description{
Synchronize selectInput with reactive
}
\keyword{internal}
