% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_SyncSelectizeInput}
\alias{srvr_SyncSelectizeInput}
\title{Synchronize selectizeInput with reactive}
\usage{
srvr_SyncSelectizeInput(
  id,
  rctv_strValue,
  chrChoices,
  ...,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_strValue}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value of a field.}

\item{chrChoices}{\code{character} Choices for a \code{\link[shiny:selectInput]{shiny::selectizeInput()}} or
other menu.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that triggers \code{\link[shiny:updateSelectInput]{shiny::updateSelectizeInput()}}
when the connected reactive changes.
}
\description{
Synchronize selectizeInput with reactive
}
\keyword{internal}
