% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_SyncTab}
\alias{srvr_SyncTab}
\title{Synchronize selected tab with reactive}
\usage{
srvr_SyncTab(
  id,
  strTargetTab,
  rctv_strValue,
  rctv_strCurrentTab,
  chrFromTabs,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strTargetTab}{\code{character} The tab to switch to.}

\item{rctv_strValue}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value of a field.}

\item{rctv_strCurrentTab}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the currently selected tab.}

\item{chrFromTabs}{\code{character} A vector of tab names. Syncing is only
triggered if the user is currently on one of these tabs.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that triggers \code{\link[bslib:nav_select]{bslib::nav_select()}}
when the connected reactive changes.
}
\description{
Synchronize selected tab with reactive
}
\keyword{internal}
