% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_SyncVirtualSelectInput}
\alias{srvr_SyncVirtualSelectInput}
\title{Synchronize virtualSelectInput with reactive}
\usage{
srvr_SyncVirtualSelectInput(
  id,
  session = getDefaultReactiveDomain(),
  rctv_chrChoices = NULL,
  rctv_strLabel = NULL,
  rctv_strSelected = NULL,
  rctv_lglDisable = NULL,
  rctv_chrDisabledChoices = NULL,
  rctv_lglOpen = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{rctv_chrChoices}{\verb{reactive character} Choices to set.}

\item{rctv_strLabel}{\verb{reactive character} Label to set.}

\item{rctv_strSelected}{\verb{reactive character} Selected value to set.}

\item{rctv_lglDisable}{\verb{reactive logical} Whether to disable the menu.}

\item{rctv_chrDisabledChoices}{\verb{reactive character} Choices to disable.}

\item{rctv_lglOpen}{\verb{reactive logical} Whether to open (\code{TRUE}) or close
(\code{FALSE}) the menu.}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that triggers
\code{\link[shinyWidgets:updateVirtualSelect]{shinyWidgets::updateVirtualSelect()}} when the connected reactives change.
}
\description{
Synchronize virtualSelectInput with reactive
}
\keyword{internal}
