% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_study_overview_server.R
\name{study_overview_server}
\alias{study_overview_server}
\title{Study Overview Server}
\usage{
study_overview_server(id, snapshot)
}
\arguments{
\item{id}{The namespace id}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}
}
\description{
Study Overview Server
}
