% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectState.R
\name{SubsetChoices}
\alias{SubsetChoices}
\title{Subset Choices}
\usage{
SubsetChoices(strDependent, l_chrChoices, envCall = rlang::caller_env())
}
\arguments{
\item{strDependent}{\code{character} The value of the dependency. Can be \code{NULL}.}

\item{l_chrChoices}{\code{list} A named list of choice vectors.}
}
\value{
\code{character} A vector of choices.
}
\description{
Subset Choices
}
\keyword{internal}
