% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_participant_input.R
\name{sync_participant_input}
\alias{sync_participant_input}
\title{Synchronize participant drop-down with click input}
\usage{
sync_participant_input(rctv_strParticipant)
}
\value{
An observer (see  \code{\link[shiny:observe]{shiny::observe()}}) that updates the participant drop-down
menu whenever the input value changes.
}
\description{
When a user clicks a participant on \code{table_participants_by_site}, also update the drop-down menu to
match.
}
\keyword{internal}
