% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_site_input.R
\name{sync_site_input}
\alias{sync_site_input}
\title{Synchronize site drop-down with click input}
\usage{
sync_site_input(rctv_strSiteID)
}
\arguments{
\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}
}
\value{
An observer (see  \code{\link[shiny:observe]{shiny::observe()}}) that updates the site drop-down
menu whenever the input value changes.
}
\description{
When a user clicks a site in a widget, also update the drop-down menu to
match.
}
\keyword{internal}
