% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglist_metric_meta_data.R
\name{taglist_metric_meta_data}
\alias{taglist_metric_meta_data}
\title{Creates taglist to summarize participants at a particular site}
\usage{
taglist_metric_meta_data(dfMetrics, strMetricID)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{strMetricID}{\code{character} A \code{MetricID} to focus on.}
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{internal}
