% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglist_site_meta_data.R
\name{taglist_site_meta_data}
\alias{taglist_site_meta_data}
\title{Creates taglist to showcase all metadata associate with a site.}
\usage{
taglist_site_meta_data(dfGroups, strSiteID)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}
}
\description{
Creates taglist to showcase all metadata associate with a site.
}
\keyword{internal}
