% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-wrangle.R
\name{temp_subsetLStudy}
\alias{temp_subsetLStudy}
\title{Subset study information}
\usage{
temp_subsetLStudy(lStudy)
}
\arguments{
\item{lStudy}{\code{list} Named list of data describing the overall study.}
}
\value{
A list with a subset of study information.
}
\description{
THIS FUNCTION SHOULD BE REMOVED AS THE APP MATURES
}
\keyword{internal}
