% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlot.R
\name{theme_gsm}
\alias{theme_gsm}
\title{Standard theme for gsm ggplot2 plots}
\usage{
theme_gsm()
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Standard theme for gsm ggplot2 plots
}
\keyword{internal}
