% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UnNS}
\alias{UnNS}
\title{Remove a namespace from an id.}
\usage{
UnNS(id, ns)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{ns}{A function to apply a namespace to an object, such as the ones
produced using \code{\link[shiny:NS]{shiny::NS()}} or the one available as \code{session$ns()} in a
Shiny server function.}
}
\value{
\code{id} minus the namespace.
}
\description{
Remove a namespace from an id.
}
\keyword{internal}
