% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metric_select.R
\name{update_metric_select}
\alias{update_metric_select}
\title{Update Metric Select}
\usage{
update_metric_select(input, output, session, snapshot)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}
}
\description{
Update Metric Select
}
