% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_participant_select.R
\name{update_participant_select}
\alias{update_participant_select}
\title{Update Participant Select}
\usage{
update_participant_select(input, output, session, snapshot)
}
\description{
Update Participant Select
}
