% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{update_sample_data}
\alias{update_sample_data}
\title{Update sample data for app development.}
\usage{
update_sample_data()
}
\description{
Update sample data for app development.
}
\examples{
\dontrun{
update_sample_data()
}

}
\keyword{internal}
