% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugins.R
\name{util_compileMissingInputs}
\alias{util_compileMissingInputs}
\title{Figure out which required inputs aren't available}
\usage{
util_compileMissingInputs(
  chrRequiredInputs,
  rctv_strSiteID,
  rctv_strSubjectID,
  rctv_strDomainID
)
}
\arguments{
\item{chrRequiredInputs}{\code{character} A vector of inputs required by a Plugin.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}
}
\value{
A character vector of missing inputs.
}
\description{
Figure out which required inputs aren't available
}
\keyword{internal}
