% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_chrDomains}
\alias{validate_chrDomains}
\title{Confirm that an object is valid chrDomains}
\usage{
validate_chrDomains(chrDomains, lPlugins = NULL, envCall = rlang::caller_env())
}
\arguments{
\item{chrDomains}{\code{character} A vector of names of domain data.frames to show
in the app. Supported domains are shown in the default value.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
\code{chrDomains} in all caps, if valid.
}
\description{
Confirm that an object is valid chrDomains
}
\keyword{internal}
