% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_dfAnalyticsInput}
\alias{validate_dfAnalyticsInput}
\title{Confirm that an object is a dfAnalyticsInput}
\usage{
validate_dfAnalyticsInput(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, if it passes the checks.
}
\description{
Confirm that an object is a dfAnalyticsInput
}
\keyword{internal}
