% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_dfBounds}
\alias{validate_dfBounds}
\title{Confirm that an object is a dfBounds}
\usage{
validate_dfBounds(x)
}
\arguments{
\item{x}{The object to validate.}
}
\value{
\code{x}, if it passes the checks.
}
\description{
Confirm that an object is a dfBounds
}
\keyword{internal}
