% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_dfGroups}
\alias{validate_dfGroups}
\title{Confirm that an object is a dfGroups}
\usage{
validate_dfGroups(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, if it passes the checks.
}
\description{
Confirm that an object is a dfGroups
}
\keyword{internal}
