% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_dfResults}
\alias{validate_dfResults}
\title{Confirm that an object is a dfResults}
\usage{
validate_dfResults(x)
}
\arguments{
\item{x}{The object to validate.}
}
\value{
\code{x}, if it passes the checks.
}
\description{
Confirm that an object is a dfResults
}
\keyword{internal}
