% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_hasOnlyFields}
\alias{validate_hasOnlyFields}
\title{Check that an object has only the allowed fields}
\usage{
validate_hasOnlyFields(
  x,
  chrAllowedFields,
  strWhat = rlang::caller_arg(x),
  envCall = rlang::caller_env()
)
}
\arguments{
\item{x}{An object to validate.}

\item{chrAllowedFields}{\code{character} A vector of names of allowed fields in an
object.}

\item{strWhat}{\code{character} A sentence-case description of the object being
inspected.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
The validated object.
}
\description{
Check that an object has only the allowed fields
}
\keyword{internal}
