% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_is_df}
\alias{validate_is_df}
\title{Confirm that an object is a data.frame}
\usage{
validate_is_df(x, strArg = rlang::caller_arg(x), envCall = rlang::caller_env())
}
\arguments{
\item{x}{The object to validate.}

\item{strArg}{\code{character} The name of the argument in the calling function.
In general, this value should either be left as the default, or passed from
the calling function if the calling function also has a \code{strArg} argument.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
\code{x}, if it is a data.frame.
}
\description{
Confirm that an object is a data.frame
}
\keyword{internal}
