% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_lPlugins}
\alias{validate_lPlugins}
\title{Confirm that an object is valid lPlugins}
\usage{
validate_lPlugins(lPlugins, envCall = rlang::caller_env())
}
\arguments{
\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
\code{lPlugins} with normalized names.
}
\description{
Confirm that an object is valid lPlugins
}
\keyword{internal}
