% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectState.R
\name{ValidateSelected}
\alias{ValidateSelected}
\title{Validate Selected Item}
\usage{
ValidateSelected(strCurrentSelection, chrChoices, strDefault = "All")
}
\arguments{
\item{strCurrentSelection}{\code{character} The current selection.}

\item{chrChoices}{\code{character} The available choices.}

\item{strDefault}{\code{character} The default value to return if the current
selection is not in the available choices.}
}
\value{
\code{character} The validated selection.
}
\description{
Validate Selected Item
}
\keyword{internal}
