% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_AETimelines.R
\name{Widget_AETimelines}
\alias{Widget_AETimelines}
\title{AE Timelines Widget}
\usage{
Widget_AETimelines(
  data,
  settings = list(id_col = "subjid", seq_col = "aeseq", stdy_col = "aest_dy", endy_col =
    "aeen_dy", term_col = "mdrpt_nsv", color = list(value_col = "aetoxgr", label =
    "Toxicity Grade"), highlight = list(value_col = "aeser", label = "Serious Event")),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{\code{data.frame} A data frame with columns:
\itemize{
\item Participant ID
\item AE Sequence Number
\item AE Start Day
\item AE End Day
\item AE Term
\item AE Severity
\item AE Seriousness
}}

\item{settings}{TODO}

\item{width}{TODO}

\item{height}{TODO}

\item{elementId}{TODO}
}
\description{
AE Timelines Widget
}
