% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_BarChart.R
\name{Widget_BarChart}
\alias{Widget_BarChart}
\title{Bar Chart Widget}
\usage{
Widget_BarChart(
  id,
  dfResults,
  strOutcome = "Score",
  lMetric = list(),
  dfGroups = NULL,
  dfBounds = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
Created with \code{\link[gsm.reporting:BindResults]{gsm.reporting::BindResults()}}.}

\item{strOutcome}{\code{character} Outcome variable. Default: \code{"Score"}.}

\item{lMetric}{\code{list} Named list of data describing a single metric, as well
as things like which group is selected.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting. Created with \code{\link[gsm.reporting:MakeBounds]{gsm.reporting::MakeBounds()}}.}
}
\description{
Bar Chart Widget
}
\keyword{internal}
