% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupOverview.R
\name{Widget_GroupOverview-shiny}
\alias{Widget_GroupOverview-shiny}
\alias{Widget_GroupOverviewOutput}
\alias{renderWidget_GroupOverview}
\title{Shiny bindings for Widget_GroupOverview}
\usage{
Widget_GroupOverviewOutput(outputId, width = "100\%", height = "100\%")

renderWidget_GroupOverview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'},
\code{'auto'}) or a number, which will be coerced to a string and have \code{'px'}
appended.}

\item{expr}{An expression that generates a Widget_GroupOverview}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{\link[=quote]{quote()}})? This is useful
if you want to save an expression in a variable.}
}
\description{
Shiny bindings for Widget_GroupOverview
}
\keyword{internal}
