% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_Plot.R
\name{Widget_Plot}
\alias{Widget_Plot}
\title{Adapt a gsm widget for use here}
\usage{
Widget_Plot(id, strWidgetName, ...)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strWidgetName}{\code{character} The name of a widget in the gsm package.}

\item{...}{Additional parameters passed on to the widget as JSON.}
}
\value{
An htmlwidget object as created by \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}.
}
\description{
This should go away eventually by making sure everything from gsm is
compatible here.
}
\keyword{internal}
