% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_scatter_plot.R
\name{widget_scatter_plot-shiny}
\alias{widget_scatter_plot-shiny}
\alias{widget_scatter_plot_output}
\alias{render_widget_scatter_plot}
\title{Shiny bindings for widget_scatter_plot}
\usage{
widget_scatter_plot_output(outputId, width = "100\%", height = "400px")

render_widget_scatter_plot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a widget_scatter_plot}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Output and render functions for using widget_scatter_plot within Shiny
applications and interactive Rmd documents.
}
