% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_scatter_plot.R
\name{widget_scatter_plot}
\alias{widget_scatter_plot}
\title{Generate Scatter Plot Widget}
\usage{
widget_scatter_plot(
  data,
  config,
  bounds,
  siteid = NULL,
  width = NULL,
  height = NULL,
  elementId = "widget_scatter_plot"
)
}
\description{
Generate Scatter Plot Widget
}
