% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TimeSeries.R
\name{Widget_TimeSeriesOutput}
\alias{Widget_TimeSeriesOutput}
\title{Shiny bindings for TimeSeries Widget}
\usage{
Widget_TimeSeriesOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'},
\code{'auto'}) or a number, which will be coerced to a string and have \code{'px'}
appended.}
}
\description{
Shiny bindings for TimeSeries Widget
}
\keyword{internal}
