% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{CheckInputUnset}
\alias{CheckInputUnset}
\title{Check whether an input is set}
\usage{
CheckInputUnset(strInputName, chrRequiredInputs, strValue)
}
\arguments{
\item{strInputName}{\code{character} The name of an input. One of \code{"group"},
\code{"participant"}, or \code{"domain"}.}

\item{chrRequiredInputs}{\code{character} An optional vector of any inputs
("metric", "group", "participant", or "domain") that must have a non-empty
value before the plugin can load. "None" and "All" count as "empty" for
this check. If the user has not set a value for that input, the app will
display a placeholder instructing the user to make a selection.}

\item{strValue}{\code{character} The value of a field.}
}
\value{
\code{strInputName} if the input is required and is not set, or
\code{character()} if it isn't required or is set.
}
\description{
Check whether an input is set
}
\keyword{internal}
