% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugins.R
\name{CheckInputUnset}
\alias{CheckInputUnset}
\title{Check whether an input is set}
\usage{
CheckInputUnset(strInputName, chrRequiredInputs, rctv_strValue)
}
\arguments{
\item{strInputName}{\code{character} The name of an input. One of \code{"site"},
\code{"participant"}, or \code{"domain"}.}

\item{chrRequiredInputs}{\code{character} A vector of inputs required by a Plugin.}

\item{rctv_strValue}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value of a field.}
}
\value{
\code{strInputName} if the input is required and is not set, or
\code{character()} if it isn't required or is set.
}
\description{
Check whether an input is set
}
\keyword{internal}
