% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Card.R
\name{choose_article}
\alias{choose_article}
\title{Choose a or an}
\usage{
choose_article(word)
}
\arguments{
\item{chrWords}{Words to choose "a" or "an" for.}
}
\value{
A vector of "a" and "an", the same length as chrWords.
}
\description{
Choose a or an
}
\keyword{internal}
