% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ColorScheme}
\alias{ColorScheme}
\title{gsm Color Scheme}
\usage{
ColorScheme(
  strColorName = c("gray", "green", "amber", "red"),
  strColorFamily = c("dark", "light")
)
}
\arguments{
\item{strColorName}{\code{character} The target color to report about.}

\item{strColorFamily}{\code{character} Whether to load the \code{"dark"} version of
this color, or the \code{"light"} version. Default: \code{"dark"}.}
}
\value{
The hex code of a color.
}
\description{
Choose a color from the "official" colors of the gsm family of packages.
}
\examples{
ColorScheme("gray")
ColorScheme("gray", "light")
ColorScheme("green")
ColorScheme("amber")
ColorScheme("red")
}
