% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{FinalWorkflowOutputNames}
\alias{FinalWorkflowOutputNames}
\title{Find the Names of the Final Steps}
\usage{
FinalWorkflowOutputNames(lWorkflows)
}
\value{
A character vector of \code{\link[=FinalWorkflowOutputName]{FinalWorkflowOutputName()}} output.
}
\description{
Find the Names of the Final Steps
}
\keyword{internal}
