% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp-package.R
\docType{package}
\name{gsm.app-package}
\alias{gsm.app}
\alias{gsm.app-package}
\title{gsm.app: gsm Shiny Application}
\description{
Explore the results of gsm assessments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-biostats.github.io/gsm.app/}
  \item \url{https://github.com/Gilead-BioStats/gsm.app}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.app/issues}
}

}
\author{
\strong{Maintainer}: Spencer Childress \email{spencer.childress@gilead.com}

Authors:
\itemize{
  \item Taylor Rodgers \email{Taylor.Rodgers@freestateanalytics.com}
  \item Matt Roumaya \email{matthewroumaya@gmail.com} (\href{https://orcid.org/0000-0002-8009-2771}{ORCID})
}

}
\keyword{internal}
