% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp-package.R
\docType{package}
\name{gsmApp-package}
\alias{gsmApp}
\alias{gsmApp-package}
\title{gsmApp: {gsm} Shiny Application}
\description{
Run {gsm} Assessments in the {safetyGraphics} Framework.
}
\author{
\strong{Maintainer}: Spencer Childress \email{spencer.childress@gilead.com}

}
\keyword{internal}
