% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp-package.R
\docType{package}
\name{gsmApp-package}
\alias{gsmApp}
\alias{gsmApp-package}
\title{gsmApp: {gsm} Shiny Application}
\description{
Run {gsm} Assessments in the {safetyGraphics} Framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://automatic-adventure-63wkll6.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Spencer Childress \email{spencer.childress@gilead.com}

Authors:
\itemize{
  \item Taylor Rodgers \email{Taylor.Rodgers@atorusresearch.com}
  \item Matt Roumaya \email{Matt.Roumaya@atorusresearch.com}
}

}
\keyword{internal}
