% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp_Server.R
\name{gsmApp_Server}
\alias{gsmApp_Server}
\title{Define Server}
\usage{
gsmApp_Server(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  fnFetchParticipantData
)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{fnFetchParticipantData}{\code{function} A function that takes a single
\code{strSubjectID} argument such as "0001", and returns a list with components
\code{metadata} and \code{metric_data}. \code{metadata} should be a named character vector
(or something that can be coerced to a character vector) of information
about the participant, and \code{metric_data} should be a named list of
data.frames, each of which contains information related to the named
metric.}
}
\value{
The main server function for use in a shiny app.
}
\description{
Define Server
}
\keyword{internal}
