% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp_UI.R
\name{gsmApp_UI}
\alias{gsmApp_UI}
\title{Shiny UI function}
\usage{
gsmApp_UI()
}
\value{
A Shiny UI object
}
\description{
This function defines the user interface for your Shiny web application.
It specifies the layout of the web page and the interface elements it contains.
}
