% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp_UI.R
\name{gsmApp_UI}
\alias{gsmApp_UI}
\title{Shiny UI function}
\usage{
gsmApp_UI(
  chrDomains,
  dfGroups,
  dfMetrics,
  dfResults,
  lPlugins = NULL,
  strTitle = ExtractAppTitle(dfGroups),
  strFavicon = "angles-up",
  strFaviconColor = ColorScheme("red"),
  tagListExtra = NULL
)
}
\arguments{
\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{strTitle}{\code{character} A title to display for the overall app.}

\item{strFavicon}{\code{character} The name of an icon to use in the browser tab
via \code{\link[favawesome:fav]{favawesome::fav()}}.}

\item{strFaviconColor}{\code{character} The hexcode or name of a color to use as
the icon fill for \code{\link[favawesome:fav]{favawesome::fav()}}.}

\item{tagListExtra}{\code{taglist} An optional \code{\link[htmltools:tagList]{htmltools::tagList()}} of
additional elements to add to the top of the app.}
}
\value{
A Shiny UI object
}
\description{
The user interface for a gsm deep dive app.
}
\keyword{internal}
