% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmApp_UI.R
\name{gsmApp_UI}
\alias{gsmApp_UI}
\title{Shiny UI function}
\usage{
gsmApp_UI(
  dfResults,
  dfMetrics,
  dfGroups,
  intNParticipants,
  strTitle = "GSM Deep Dive"
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{intNParticipants}{\code{integer} The number of unique participants
(subjects) in the study.}

\item{strTitle}{\code{character} A title to display for the overall app.}
}
\value{
A Shiny UI object
}
\description{
This function defines the user interface for your Shiny web application.
It specifies the layout of the web page and the interface elements it contains.
}
