% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-widgets.R
\name{gsmDependencies}
\alias{gsmDependencies}
\title{Get dependencies for a gsm widget}
\usage{
gsmDependencies(strWidgetName, excludes = character())
}
\arguments{
\item{strWidgetName}{The name of a widget in the gsm package.}

\item{excludes}{An optional vector of named dependencies to \emph{not} include.}
}
\value{
A list of \code{htmltools::htmlDependency} objects for use in
\code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}.
}
\description{
Get dependencies for a gsm widget
}
\keyword{internal}
