% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_GroupOverview.R
\name{htmlDependency_GroupOverviewInput}
\alias{htmlDependency_GroupOverviewInput}
\title{Group Overview Input JavaScript}
\usage{
htmlDependency_GroupOverviewInput()
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} of \code{html_dependency} objects (see
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}), so that each will be attached to the Shiny
app exactly once, regardless how many times they are added.
}
\description{
Attach \code{groupOverviewInput.js} to an app or other HTML exactly once.
}
\keyword{internal}
