% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_WidgetPlot.R
\name{htmlDependency_WidgetInputHelpers}
\alias{htmlDependency_WidgetInputHelpers}
\title{Widget Input JavaScript}
\usage{
htmlDependency_WidgetInputHelpers()
}
\value{
An \code{html_dependency} object (see \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}),
which is attached to the Shiny app exactly once, regardless how many times
it is added.
}
\description{
Attach \code{widgetInputHelpers.js} to an app or other HTML exactly once.
}
\keyword{internal}
