% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_metric_select.R
\name{initialize_metric_select}
\alias{initialize_metric_select}
\title{Initialize Metric Select}
\usage{
initialize_metric_select(dfMetrics, session)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{session}{Shiny session}
}
\description{
Initialize Metric Select
}
