% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_metric_select.R
\name{initialize_metric_select}
\alias{initialize_metric_select}
\title{Initialize Metric Select}
\usage{
initialize_metric_select(dfMetrics, session)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\description{
Initialize Metric Select
}
