% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_participant_select.R
\name{initialize_participant_select}
\alias{initialize_participant_select}
\title{Initialize Participant Select}
\usage{
initialize_participant_select(dfAnalyticsInput, session)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\description{
Initialize Participant Select
}
