% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_site_select.R
\name{initialize_site_select}
\alias{initialize_site_select}
\title{Initialize Site Select}
\usage{
initialize_site_select(input, output, session, snapshot)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}
}
\description{
Initialize Site Select
}
