% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_site_select.R
\name{initialize_site_select}
\alias{initialize_site_select}
\title{Initialize Site Select}
\usage{
initialize_site_select(dfGroups, session)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\description{
Initialize Site Select
}
\keyword{internal}
