% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_character.R
\name{input_character}
\alias{input_character}
\title{Generates character input}
\usage{
input_character(id, default)
}
\arguments{
\item{id}{\code{character} Input ID}

\item{default}{\code{character} Input metadata}
}
\description{
Generates character input
}
