% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata.R
\name{map_metadata}
\alias{map_metadata}
\title{Structure {gsm} Metadata}
\usage{
map_metadata(meta = clindata::mapping_rawplus)
}
\arguments{
\item{meta}{\code{list} clindata::mapping_rawplus}
}
\value{
\code{data.frame} {gsm} metadata in tabular form
}
\description{
Map {gsm} metadata to the tabular structure expected by {safetyGraphics}
}
\examples{
metadata <- map_metadata()

}
