% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ActionList.R
\name{mod_ActionList_Server}
\alias{mod_ActionList_Server}
\title{List of ActionLinks Server}
\usage{
mod_ActionList_Server(id, chrLinkIDs)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrLinkIDs}{\code{character} Module ids for multiple modules.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} with the id of the most recently clicked
link.
}
\description{
The server component for \code{\link[=mod_ActionList_UI]{mod_ActionList_UI()}}. This server keeps track of
which link was most recently clicked.
}
