% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ActionList.R
\name{mod_ActionList_UI}
\alias{mod_ActionList_UI}
\title{List of ActionLinks UI}
\usage{
mod_ActionList_UI(id, chrLinkIDs, chrLabels, chrValues)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrLinkIDs}{\code{character} Module ids for multiple modules.}

\item{chrLabels}{\code{character} A vector of labels for display to the user.}

\item{chrValues}{\code{character} A vector of values to associate with a vector of
labels.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} of labels and values, where the labels are
\code{\link[shiny:actionButton]{shiny::actionLink()}} items.
}
\description{
Display a vertical list of \code{\link[shiny:actionButton]{shiny::actionLink()}} items, along with a
corresponding vertical list of values associated with those links.
}
\examples{
mod_ActionList_UI(
  "demo",
  c("link1", "link2", "link3"),
  c("Label 1", "Label 2", "Label 3"),
  c("value 1", "value 2", "value 3")
)
}
