% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ActionList_UI.R
\name{mod_ActionList_UI}
\alias{mod_ActionList_UI}
\title{List of ActionLinks UI}
\usage{
mod_ActionList_UI(id, chrLinkIDs, chrLabels, chrValues)
}
\arguments{
\item{id}{The id for this module in this namespace.}

\item{chrLinkIDs}{\code{character} Module ids for multiple modules.}

\item{chrLabels}{\code{character} A vector of labels for display to the user.}

\item{chrValues}{\code{character} A vector of values to associate with a vector of
labels.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} of labels and values, where the labels are
\code{\link[shiny:actionButton]{shiny::actionLink()}} items.
}
\description{
List of ActionLinks UI
}
\keyword{internal}
