% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CascadingSelect.R
\name{mod_CascadingSelect_UI}
\alias{mod_CascadingSelect_UI}
\title{Cascading Select Input UI}
\usage{
mod_CascadingSelect_UI(id, strLabel, chrChoices = "All")
}
\arguments{
\item{id}{\code{character} The module ID.}

\item{strLabel}{\code{character} The initial label for the select input.}

\item{chrChoices}{\code{character} The initial set of choices for the select
input.}
}
\value{
A \code{bslib::nav_item()}
}
\description{
Cascading Select Input UI
}
\keyword{internal}
