% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainData.R
\name{mod_DomainData_Server}
\alias{mod_DomainData_Server}
\title{Domain Data Server}
\usage{
mod_DomainData_Server(id, rctv_dfDomain)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the SubjectID for each row that is
selected. Currently not useful but will be converted to row numbers when we
need it for deeper dives.
}
\description{
Display a specific (named) table, or a placeholder.
}
\keyword{internal}
