% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainData_Server.R
\name{mod_DomainData_Server}
\alias{mod_DomainData_Server}
\title{Domain Data Server}
\usage{
mod_DomainData_Server(id, rctv_lData, rctv_strSubjectID)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_lData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns a
named list of dataframes.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the SubjectID for each row that is
selected. Currently not useful but will be converted to row numbers when we
need it for deeper dives.
}
\description{
Display a specific (named) table, or a placeholder.
}
\keyword{internal}
