% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainData.R
\name{mod_DomainData_UI}
\alias{mod_DomainData_UI}
\title{Domain Data Tab UI}
\usage{
mod_DomainData_UI(id, strDomain)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strDomain}{\code{character} The domain data.frame to load. Supported domains
are shown in the default value.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} with either a placeholder, or a
\code{\link[gt:gt_output]{gt::gt_output()}} table.
}
\description{
Domain Data Tab UI
}
\keyword{internal}
