% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainDetails.R
\name{mod_DomainDetails_Server}
\alias{mod_DomainDetails_Server}
\title{Domain Details server}
\usage{
mod_DomainDetails_Server(id, l_rctvDomains, rctv_strDomainID)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}
}
\description{
Update Domain Details when the selected participant changes.
}
\keyword{internal}
