% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainDetails_Server.R
\name{mod_DomainDetails_Server}
\alias{mod_DomainDetails_Server}
\title{Domain Details server}
\usage{
mod_DomainDetails_Server(
  id,
  fnFetchData,
  chrDomains = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY"),
  rctv_strSiteID,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomain} argument
and optional \code{strSiteID} and \code{strSubjectID} and returns a data.frame. See
\code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The returned data.frame contains
information about the named domain.}

\item{chrDomains}{\code{character} A vector of names of domain data.frames to show
in the app. Supported domains are shown in the default value.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the ids of rows selected in the table
(currently unused but implemented as a proof-of-concept for future
extensions).
}
\description{
Update Domain Details when the selected participant changes.
}
\keyword{internal}
