% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainDetails_Server.R
\name{mod_DomainDetails_Server}
\alias{mod_DomainDetails_Server}
\title{Domain Details server}
\usage{
mod_DomainDetails_Server(id, l_rctvDomains)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the ids of rows selected in the table
(currently unused but implemented as a proof-of-concept for future
extensions).
}
\description{
Update Domain Details when the selected participant changes.
}
\keyword{internal}
