% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainDetails.R
\name{mod_DomainDetails_UI}
\alias{mod_DomainDetails_UI}
\title{Domain Details UI}
\usage{
mod_DomainDetails_UI(
  id,
  chrDomains = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY")
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrDomains}{\code{character} A vector of names of domain data.frames to show
in the app. Supported domains are shown in the default value.}
}
\value{
A \code{\link[bslib:navset]{bslib::navset_underline()}} with a tab for each available domain.
}
\description{
Domain Details UI
}
\keyword{internal}
