% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainCount.R
\name{mod_DomainLabel_UI}
\alias{mod_DomainLabel_UI}
\title{Domain Tab Label UI}
\usage{
mod_DomainLabel_UI(id, strDomainLabel, strDomainID)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{strDomainLabel}{\code{character} The domain name to display. For example,
for ID \code{"AE"}, the label is \code{"Adverse Events"}. Can also be a
\code{\link[shiny:htmlOutput]{shiny::uiOutput()}} which renders to an inline string.}

\item{strDomainID}{\code{character} The domain data.frame to load.}
}
\value{
A \code{\link[shiny:reexports]{shiny::HTML()}} that includes a dynamic \code{\link[shiny:htmlOutput]{shiny::uiOutput()}} with
the count of rows in this domain.
}
\description{
Domain Tab Label UI
}
\keyword{internal}
