% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DomainSummary.R
\name{mod_DomainSummary_Server}
\alias{mod_DomainSummary_Server}
\title{Dynamically Generate the Domain Summary}
\usage{
mod_DomainSummary_Server(
  id,
  rctv_strDomainID,
  rctv_intDomainCounts,
  chrDomains
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}

\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}
}
\description{
Dynamically Generate the Domain Summary
}
\keyword{internal}
