% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DynamicLabelKey.R
\name{mod_DynamicLabelKey_UI}
\alias{mod_DynamicLabelKey_UI}
\title{Study, Group, and Participant label buttons}
\usage{
mod_DynamicLabelKey_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[htmltools:builder]{htmltools::span()}} element containing a set of
\code{\link[shinyWidgets:checkboxGroupButtons]{shinyWidgets::checkboxGroupButtons()}}.
}
\description{
Study, Group, and Participant label buttons
}
\keyword{internal}
