% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupDetails.R
\name{mod_GroupDetails_Server}
\alias{mod_GroupDetails_Server}
\title{Group Details Server}
\usage{
mod_GroupDetails_Server(
  id,
  dfGroups,
  dfAnalyticsInput,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  rctv_strMetricID,
  rctv_lMetric
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data. This
data.frame is created by binding together the \code{lAnalysis} "Analysis_Input"
list of data.frames output of analysis workflows, with a \code{MetricID} column
for the name of each list.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the id of the most recently selected
participant.
}
\description{
Group Details Server
}
\keyword{internal}
