% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupDetails.R
\name{mod_GroupDetails_UI}
\alias{mod_GroupDetails_UI}
\title{Group Details UI}
\usage{
mod_GroupDetails_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:layout_columns]{bslib::layout_columns()}} with group metadata and group participants.
}
\description{
Group Details UI
}
\keyword{internal}
