% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupInput.R
\name{mod_GroupInput_UI}
\alias{mod_GroupInput_UI}
\title{Group and GroupLevel Dropdown Menu UI}
\usage{
mod_GroupInput_UI(id, dfGroups)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}
}
\value{
A \code{list} with one or two UI elements.
}
\description{
Group and GroupLevel Dropdown Menu UI
}
\keyword{internal}
