% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupOverview.R
\name{mod_GroupOverview_Server}
\alias{mod_GroupOverview_Server}
\title{Group Overview Module Server}
\usage{
mod_GroupOverview_Server(
  id,
  dfResults,
  dfMetrics,
  dfGroups,
  rctv_strGroupID,
  rctv_strMetricID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}
}
\description{
Group Overview Module Server
}
\keyword{internal}
