% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupOverview_Server.R
\name{mod_GroupOverview_Server}
\alias{mod_GroupOverview_Server}
\title{Group Overview Module Server}
\usage{
mod_GroupOverview_Server(id, dfResults, dfMetrics, dfGroups)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}
}
\value{
A list of two \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} values: \code{rctv_strGroupID} (the
selected group, if any) and \code{rctv_strMetricID} (the selected metric, if
any).
}
\description{
Group Overview Module Server
}
\keyword{internal}
