% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_GroupParticipants.R
\name{mod_GroupParticipants_Server}
\alias{mod_GroupParticipants_Server}
\title{Group Participants Server}
\usage{
mod_GroupParticipants_Server(
  id,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strSubjectID,
  rctv_dfAnalyticsInput,
  rctv_lColumnNames
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{rctv_lColumnNames}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a named list of column names to substitute into tables for display.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} with the id of the participant selected in
the table.
}
\description{
Group Participants Server
}
\keyword{internal}
