% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metric_details_server.R
\name{mod_metric_details_server}
\alias{mod_metric_details_server}
\title{Metric Details Server}
\usage{
mod_metric_details_server(
  id,
  dfResults,
  dfMetrics,
  dfGroups,
  dfBounds,
  rctv_strSiteID,
  rctv_strMetricID
)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{rctv_strSiteID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strMetricID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the selected \code{MetricID}.}
}
\description{
Metric Details Server
}
