% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricDetails_Server.R
\name{mod_MetricDetails_Server}
\alias{mod_MetricDetails_Server}
\title{Metric Details Server}
\usage{
mod_MetricDetails_Server(
  id,
  dfResults,
  dfGroups,
  dfBounds,
  rctv_lMetric,
  rctv_strSiteID,
  rctv_strMetricID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the id of the group selected that was
most recently selected on one of the tabs.
}
\description{
Metric Details Server
}
\keyword{internal}
