% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricDetails.R
\name{mod_MetricDetails_UI}
\alias{mod_MetricDetails_UI}
\title{Metric Details UI}
\usage{
mod_MetricDetails_UI(id, dfMetrics)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}
}
\description{
Metric Details UI
}
\keyword{internal}
