% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricDetails.R
\name{mod_MetricDetails_UI}
\alias{mod_MetricDetails_UI}
\title{Metric Details UI}
\usage{
mod_MetricDetails_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\description{
Metric Details UI
}
\keyword{internal}
