% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MetricTable_Server.R
\name{mod_MetricTable_Server}
\alias{mod_MetricTable_Server}
\title{Metric Table Module Server}
\usage{
mod_MetricTable_Server(id, rctv_dfResults, dfGroups, rctv_strSiteID)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_dfResults}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a stacked summary of analysis pipeline output.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the id of the selected group.
}
\description{
Metric Table Module Server
}
\keyword{internal}
