% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_Server.R
\name{mod_ParticipantDetails_Server}
\alias{mod_ParticipantDetails_Server}
\title{Participant Details server}
\usage{
mod_ParticipantDetails_Server(id, fnFetchData, rctv_strSubjectID)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomain} argument
and optional \code{strSiteID} and \code{strSubjectID} and returns a data.frame. See
\code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The returned data.frame contains
information about the named domain.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the ids of rows selected in the table
(currently unused but implemented as a proof-of-concept for future
extensions).
}
\description{
Update Participant Details when the selected participant changes.
}
\keyword{internal}
