% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDetails_UI.R
\name{mod_ParticipantDetails_UI}
\alias{mod_ParticipantDetails_UI}
\title{Participant Details UI}
\usage{
mod_ParticipantDetails_UI(id)
}
\arguments{
\item{id}{The id for this module in this namespace.}
}
\description{
Participant Details UI
}
\keyword{internal}
