% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantDomain_UI.R
\name{mod_ParticipantDomain_UI}
\alias{mod_ParticipantDomain_UI}
\title{Participant Domain Details Card UI}
\usage{
mod_ParticipantDomain_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with either a placeholder, or a
\code{\link[DT:dataTableOutput]{DT::DTOutput()}} table.
}
\description{
Participant Domain Details Card UI
}
\keyword{internal}
