% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ParticipantMetricSummary_Server.R
\name{mod_ParticipantMetricSummary_Server}
\alias{mod_ParticipantMetricSummary_Server}
\title{Participant Metric Summary server}
\usage{
mod_ParticipantMetricSummary_Server(id, rctv_lParticipantMetricData)
}
\value{
A string identifying the selected summary (currently does not work
and instead returns the string "selectedID") as a \code{\link[shiny:reactive]{shiny::reactive()}}.
}
\description{
Inputs as links to metric summary reports, with counts of rows.
}
\keyword{internal}
