% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugin.R
\name{mod_Plugin_UI}
\alias{mod_Plugin_UI}
\title{Plugin Wrapper UI}
\usage{
mod_Plugin_UI(lPlugin, id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
Plugin UI element, as a \code{\link[shiny:htmlOutput]{shiny::uiOutput()}}. The corresponding
server function determines whether the plugin is ready to display, showing
either a placeholder or the actual plugin.
}
\description{
Plugin Wrapper UI
}
\keyword{internal}
