% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugins_Server.R
\name{mod_Plugins_Server}
\alias{mod_Plugins_Server}
\title{Plugins Wrapper Server}
\usage{
mod_Plugins_Server(
  id,
  lPlugins,
  l_rctvDomains,
  rctv_strMetricID,
  rctv_strSiteID,
  rctv_strSubjectID,
  rctv_strDomainID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}
}
\value{
Nothing yet.
}
\description{
Plugins Wrapper Server
}
\keyword{internal}
