% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugins_Server.R
\name{mod_Plugins_Server}
\alias{mod_Plugins_Server}
\title{Plugins Wrapper Server}
\usage{
mod_Plugins_Server(
  id,
  lPlugins,
  fnFetchData,
  rctv_strMetricID,
  rctv_strSiteID,
  rctv_strSubjectID
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomain} argument
and optional \code{strSiteID} and \code{strSubjectID} and returns a data.frame. See
\code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The returned data.frame contains
information about the named domain.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
Nothing yet.
}
\description{
Plugins Wrapper Server
}
\keyword{internal}
