% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PotentialChoice.R
\name{mod_PotentialChoice_UI}
\alias{mod_PotentialChoice_UI}
\title{Potential Choice Select Input UI}
\usage{
mod_PotentialChoice_UI(id, chrChoices, strLabel)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrChoices}{\code{character} A (named, optionally) vector of choices for a
\code{\link[shiny:selectInput]{shiny::selectInput()}} or similar input function.}

\item{strLabel}{\code{character} The label of a field.}
}
\value{
An \code{\link[=input_NavbarExtra]{input_NavbarExtra()}} element or \code{NULL}.
}
\description{
Renders a select input only if there is more than one choice.
}
\keyword{internal}
