% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_PrevalencePlot.R
\name{mod_PrevalencePlot_UI}
\alias{mod_PrevalencePlot_UI}
\title{Plot part of the prevalence plot UI}
\usage{
mod_PrevalencePlot_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[gsm.vizr:barsOutput]{gsm.vizr::barsOutput()}} object.
}
\description{
Plot part of the prevalence plot UI
}
\keyword{internal}
