% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_RAGPillSet.R
\name{mod_RAGPillSet_Server}
\alias{mod_RAGPillSet_Server}
\title{Set of RAG Toggle Pills Server}
\usage{
mod_RAGPillSet_Server(id, rctv_strGroupSubset = reactive("red"))
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_strGroupSubset}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected subset of groups to include in the table.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the state of the selection as a string.
}
\description{
Set of RAG Toggle Pills Server
}
\keyword{internal}
