% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlot_Server.R
\name{mod_ScatterPlot_Server}
\alias{mod_ScatterPlot_Server}
\title{ScatterPlot Module Server}
\usage{
mod_ScatterPlot_Server(
  id,
  rctv_dfResults,
  rctv_lMetric,
  dfGroups,
  rctv_dfBounds
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{rctv_dfResults}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a stacked summary of analysis pipeline output.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{rctv_dfBounds}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a set of predicted percentages/rates and upper- and lower-bounds
across the full range of sample sizes/total exposure values for reporting.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} with the id of the selected group.
}
\description{
ScatterPlot Module Server
}
\keyword{internal}
