% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlot_UI.R
\name{mod_ScatterPlot_UI}
\alias{mod_ScatterPlot_UI}
\title{ScatterPlot Module UI}
\usage{
mod_ScatterPlot_UI(id, strPlotTitle = NULL)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and a
\code{\link[=Widget_ScatterPlotOutput]{Widget_ScatterPlotOutput()}}.
}
\description{
ScatterPlot Module UI
}
\keyword{internal}
