% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlotSet.R
\name{mod_ScatterPlotSet_Server}
\alias{mod_ScatterPlotSet_Server}
\title{ScatterPlotSet Module Server}
\usage{
mod_ScatterPlotSet_Server(
  id,
  dfResults,
  dfMetrics,
  dfGroups,
  dfBounds,
  rctv_strGroupID,
  rctv_strGroupLevel,
  rctv_strMetricID
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{rctv_strGroupID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected group (usually site), and can be
used to update which group is selected.}

\item{rctv_strGroupLevel}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}
object that returns the selected \code{GroupLevel}.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}
}
\description{
ScatterPlotSet Module Server
}
\keyword{internal}
