% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ScatterPlotSet_UI.R
\name{mod_ScatterPlotSet_UI}
\alias{mod_ScatterPlotSet_UI}
\title{ScatterPlot Set Module UI}
\usage{
mod_ScatterPlotSet_UI(id, chrMetrics)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}
}
\value{
An \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \code{\link[bslib:layout_column_wrap]{bslib::layout_column_wrap()}}
of plots, and the output of \code{\link[=out_DetectCardClicks]{out_DetectCardClicks()}} (the JavaScript
necessary to detect that the plot has been clicked).
}
\description{
ScatterPlot Set Module UI
}
\keyword{internal}
