% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectDataCategory.R
\name{mod_SelectDataCategory_UI}
\alias{mod_SelectDataCategory_UI}
\title{A pass-through for a type of inline select}
\usage{
mod_SelectDataCategory_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[htmltools:builder]{htmltools::span()}} element containing necessary elements for an
inline \code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}}.
}
\description{
A pass-through for a type of inline select
}
\keyword{internal}
