% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteDetails_Server.R
\name{mod_SiteDetails_Server}
\alias{mod_SiteDetails_Server}
\title{Site Details Server}
\usage{
mod_SiteDetails_Server(
  id,
  dfGroups,
  dfAnalyticsInput,
  rctv_strSiteID,
  rctv_strMetricID,
  rctv_lMetric
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}
}
\description{
Site Details Server
}
\keyword{internal}
