% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SiteDetails.R
\name{mod_SiteDetails_UI}
\alias{mod_SiteDetails_UI}
\title{Site Details UI}
\usage{
mod_SiteDetails_UI(id)
}
\arguments{
\item{id}{\code{character} The id for this element.}
}
\value{
A \code{\link[bslib:layout_columns]{bslib::layout_columns()}} with site metadata and site participants.
}
\description{
Site Details UI
}
\keyword{internal}
