% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_StudyOverview_Server.R
\name{mod_StudyOverview_Server}
\alias{mod_StudyOverview_Server}
\title{Study Overview Server}
\usage{
mod_StudyOverview_Server(id, dfResults, dfMetrics, dfGroups)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}
}
\description{
Study Overview Server
}
\keyword{internal}
