% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_StudyOverview.R
\name{mod_StudyOverview_UI}
\alias{mod_StudyOverview_UI}
\title{Study Overview UI}
\usage{
mod_StudyOverview_UI(id, dfGroups, dfResults, chrMetrics)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}
}
\description{
Study Overview UI
}
\keyword{internal}
