% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_StudyOverview_UI.R
\name{mod_StudyOverview_UI}
\alias{mod_StudyOverview_UI}
\title{Study Overview UI}
\usage{
mod_StudyOverview_UI(id)
}
\arguments{
\item{id}{The id for this module in this namespace.}
}
\description{
Study Overview UI
}
\keyword{internal}
