% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TogglePill_Server.R
\name{mod_TogglePill_Server}
\alias{mod_TogglePill_Server}
\title{Toggle Pill Module Server}
\usage{
mod_TogglePill_Server(id, rctv_lglState = shiny::reactive(FALSE))
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{rctv_lglState}{\verb{reactive logical} A \code{\link[shiny:reactive]{shiny::reactive()}}] object that
returns a Boolean value indicating whether something is "off" (\code{FALSE}) or
"on" (\code{TRUE}).}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} indicating whether the pill is "on"
(\code{TRUE}) or "off" (\code{FALSE}).
}
\description{
Toggle Pill Module Server
}
\keyword{internal}
