% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_WidgetPlot_UI.R
\name{mod_WidgetPlot_UI}
\alias{mod_WidgetPlot_UI}
\title{Widget Plot Module UI}
\usage{
mod_WidgetPlot_UI(id, fnHtmlDependency, fnWidgetOutput, strPlotTitle = NULL)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with an optional title and a
\code{\link[=Widget_BarChartOutput]{Widget_BarChartOutput()}}.
}
\description{
Widget Plot Module UI
}
\keyword{internal}
