% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_metric_select.R
\name{observe_metric_select}
\alias{observe_metric_select}
\title{Observe Metric Select}
\usage{
observe_metric_select(snapshot, metric)
}
\arguments{
\item{snapshot}{The snapshot \code{list} object that is passed on from \code{run_app()}}

\item{metric}{A reactive value of the metric input found in the \code{ui} function.}
}
\description{
Observe Metric Select
}
