% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Card.R
\name{out_Card}
\alias{out_Card}
\title{Standardized card}
\usage{
out_Card(tagTitle, ..., id = NULL)
}
\arguments{
\item{tagTitle}{\verb{html tag} A string or other html to use as the title of
the card. Will be wrapped inside \code{\link[bslib:card_body]{bslib::card_title()}}.}

\item{...}{\verb{html tags} Objects to place in the card.}

\item{id}{\code{character} The id for this module, widget, or other element.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}}.
}
\description{
Standardized card
}
\keyword{internal}
