% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Card.R
\name{out_CardSubtitle}
\alias{out_CardSubtitle}
\title{Card subtitle}
\usage{
out_CardSubtitle(tagTitle, chrAdditionalClass = character())
}
\arguments{
\item{tagTitle}{\verb{html tag} A string or other html to use as the title of
the card. Will be wrapped inside \code{\link[bslib:card_body]{bslib::card_title()}}.}

\item{chrAdditionalClass}{\code{character} One or more additional classes to apply
to the subtitle.}
}
\value{
A \code{\link[bslib:card_body]{bslib::card_title()}} with appropriate classes.
}
\description{
Card subtitle
}
\keyword{internal}
