% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_MainTabs.R
\name{out_MainTabs}
\alias{out_MainTabs}
\title{Main Tabbed Content}
\usage{
out_MainTabs(dfResults, chrMetrics)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}
}
\value{
A list of \code{\link[bslib:nav-items]{bslib::nav_panel()}} elements, containing the main tabbed
content.
}
\description{
Main Tabbed Content
}
\keyword{internal}
