% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_NavbarExtras.R
\name{out_NavbarExtras}
\alias{out_NavbarExtras}
\title{Things on the right side of navbar}
\usage{
out_NavbarExtras(dfGroups)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary. See
\code{vignette("DataReporting", package = "gsm.reporting")} for an example.}
}
\value{
A list of \code{\link[bslib:nav_item]{bslib::nav_item()}} with
\code{\link[shinyWidgets:virtualSelectInput]{shinyWidgets::virtualSelectInput()}}.
}
\description{
Things on the right side of navbar
}
\keyword{internal}
