% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_NavbarExtras.R
\name{out_NavbarExtras}
\alias{out_NavbarExtras}
\title{Things on the right side of navbar}
\usage{
out_NavbarExtras(chrSites, intNParticipants)
}
\arguments{
\item{chrSites}{\code{character} A vector of sites available in the study.}

\item{intNParticipants}{\code{integer} The number of unique participants
(subjects) in the study.}
}
\value{
A \code{\link[htmltools:tagList]{htmltools::tagList()}} with app inputs.
}
\description{
Things on the right side of navbar
}
\keyword{internal}
