% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantMetadata.R
\name{out_ParticipantMetadata}
\alias{out_ParticipantMetadata}
\title{Participant metadata card}
\usage{
out_ParticipantMetadata(lParticipantMetadata)
}
\arguments{
\item{lParticipantMetadata}{\code{list} Named list of data describing a single
participant.}
}
\value{
An \code{\link[htmltools:builder]{htmltools::div()}} with participant metadata.
}
\description{
Participant metadata card
}
\keyword{internal}
