% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_ParticipantsBySite.R
\name{out_ParticipantsBySite}
\alias{out_ParticipantsBySite}
\title{Table of Participants Based on Site Selection}
\usage{
out_ParticipantsBySite(dfAnalyticsInput, dfMetrics, strSite, strMetricID)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{strSite}{\code{character} The \code{GroupID} of a site to focus on.}

\item{strMetricID}{\code{character} A \code{MetricID} to focus on.}
}
\description{
Table of Participants Based on Site Selection
}
\keyword{internal}
