% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Card.R
\name{out_Placeholder}
\alias{out_Placeholder}
\title{Placeholder card}
\usage{
out_Placeholder(chrRequirement)
}
\arguments{
\item{chrRequirement}{\code{character} Text to describe what is required, as a
string or character vector.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with a placeholder.
}
\description{
Placeholder card
}
\keyword{internal}
