% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_Sidebar.R
\name{out_Sidebar}
\alias{out_Sidebar}
\title{Study Information Sidebar}
\usage{
out_Sidebar(
  lStudy,
  chrMetrics,
  chrSites,
  intNParticipants,
  tagListSidebar = NULL
)
}
\arguments{
\item{lStudy}{\code{list} Named list of data describing the overall study.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}

\item{chrSites}{\code{character} A vector of sites available in the study.}

\item{intNParticipants}{\code{integer} The number of unique participants
(subjects) in the study.}

\item{tagListSidebar}{\code{taglist} An optional \code{\link[htmltools:tagList]{htmltools::tagList()}} of
additional elements to add to the top of the app sidebar.}
}
\value{
A \code{\link[bslib:sidebar]{bslib::sidebar()}} with study information, inputs, and the
stylesheets and dependencies required to display the app.
}
\description{
Study Information Sidebar
}
\keyword{internal}
