% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_StudyInformation.R
\name{out_StudyInformation}
\alias{out_StudyInformation}
\title{Study Information Card}
\usage{
out_StudyInformation(dfGroups, dfResults)
}
\arguments{
\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} with overall study metadata.
}
\description{
Study Information Card
}
\keyword{internal}
