% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant_details_server.R
\name{participant_details_server}
\alias{participant_details_server}
\title{Participant Details Server}
\usage{
participant_details_server(id, snapshot, participant)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}

\item{participant}{The reactive value provided by the participant input from \code{server}}
}
\description{
Participant Details Server
}
