% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant_details_server.R
\name{participant_details_server}
\alias{participant_details_server}
\title{Participant Details Server}
\usage{
participant_details_server(id, snapshot, participant)
}
\arguments{
\item{id}{The namespace id}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}

\item{participant}{The reactive value provided by the participant input from \code{server}}
}
\description{
Participant Details Server
}
