% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_metric_summary_tag_list.R
\name{participant_metric_summary_tag_list}
\alias{participant_metric_summary_tag_list}
\title{Creates taglist to summarize participants at a particular site}
\usage{
participant_metric_summary_tag_list(id, participant, snapshot)
}
\arguments{
\item{participant_data}{A combined data set with the required metrics.}
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{'Enrollment',}
\keyword{'Queries'}
\keyword{=}
\keyword{dfENROLL'}
\keyword{dfQUERY'}
\keyword{internal}
