% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_InstallDependencySources}
\alias{plugin_InstallDependencySources}
\title{Install Plugin Package Dependencies}
\usage{
plugin_InstallDependencySources(lPlugin)
}
\arguments{
\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
\code{lPlugin}, invisibly. This function is called for its side effects.
}
\description{
Plugins can have additional dependencies that are invisible to the usual
CI/CD pipelines. Use this function to install any such dependencies using the
pak package.
}
