% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_LoadDependencies}
\alias{plugin_LoadDependencies}
\title{Load Plugin Dependencies}
\usage{
plugin_LoadDependencies(lPluginDefinition)
}
\arguments{
\item{lPluginDefinition}{\code{list} A named list with required elements
\code{strTitle}, \code{fnUI}, and \code{fnServer}, and optional fields \code{packages} and
\code{lConfig}. Usually generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
\code{lPluginDefinition}, invisibly. This function is called for its side
effects.
}
\description{
Load the package dependencies of a plugin. This is designed to be used in an
\code{app.R} file to ensure that the dependencies are detected by packages like
rsconnect.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
plugin_LoadDependencies(list(packages = list(list(name = "gsm.app"))))
\dontshow{\}) # examplesIf}
}
