% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_Read.R
\name{plugin_Read}
\alias{plugin_Read}
\title{Read a Plugin Definition}
\usage{
plugin_Read(strPath)
}
\arguments{
\item{strPath}{The directory that contains the plugin.}
}
\value{
A \code{list} with elements \code{meta}, \code{shiny}, \code{domains}, and (optionally)
\code{required_inputs} and/or \code{packages}, read from the YAML file in \code{strPath}.
As a side effect, any \code{R} files in \code{strPath} are also loaded using
\code{source()}.
}
\description{
Plugins are defined by a named \code{list} with elements \code{meta} (containing at
least a \code{Name} to show in the plugin menu), \code{shiny} (containing \code{UI} and
\code{Server} fields to define the Shiny UI and Server functions for the plugin),
and \code{domains} (to specify the data domains that the plugin uses), plus
optional fields \code{required_inputs} and \code{packages}. Use \code{plugin_Read()} to read
these element definitions and any R files in the same directory.
}
\examples{
subjPlugin <- plugin_Read(
  system.file("plugins", "ParticipantProfile", package = "gsm.app")
)
subjPlugin
}
