% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_assessment.R
\name{prepare_assessment}
\alias{prepare_assessment}
\title{Prepare {gsm} Assessment for Ingestion by {safetyGraphics}}
\usage{
prepare_assessment(
  domain,
  yaml_file = system.file("assessments", paste0(tolower(domain), ".yaml"), package =
    "gsmApp"),
  meta = map_metadata(),
  workflows = get_metadata_from_yaml("workflows"),
  assessment = list(map_function = NULL, assess_function = NULL)
)
}
\arguments{
\item{domain}{\code{character} name of data domain associated with assessment}

\item{yaml_file}{\code{character} file path of .yaml file containing assessment metadata}

\item{meta}{\code{data.frame} metadata returned by \code{\link[=map_metadata]{map_metadata()}}}

\item{workflows}{\code{data.frame} metadata returned by \code{\link[=map_metadata]{map_metadata()}}}
}
\description{
Define a chart object with the format expected by {safetyGraphics}, including required data
domains, the charting function, and relevnat metadata.
}
