% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run App}
\usage{
run_app(snapshot = NULL)
}
\arguments{
\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Run App
}
