% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsm_app.R
\name{run_gsm_app}
\alias{run_gsm_app}
\title{Run App}
\usage{
run_gsm_app(
  dfResults,
  dfGroups,
  dfMetrics,
  dfBounds,
  dfAnalyticsInput,
  fnFetchParticipantData,
  strTitle = "GSM Deep Dive"
)
}
\arguments{
\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{fnFetchParticipantData}{\code{function} A function that takes a single
\code{strSubjectID} argument such as "0001", and returns a list with components
\code{metadata} and \code{metric_data}. \code{metadata} should be a named charact vector
(or something that can be coerced to a character vector) of information
about the participant, and \code{metric_data} should be a named list of
data.frames, each of which contains information related to the named
metric.}

\item{strTitle}{\code{character} A title to display for the overall app.}
}
\description{
Run App
}
