% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Plugin.R
\name{run_PluginUI}
\alias{run_PluginUI}
\title{Run Plugin UI}
\usage{
run_PluginUI(id, lPlugin)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{lPlugin}{\code{list} A named list with required elements \code{meta} and \code{shiny},
and optional fields \code{spec}, \code{packages} and \code{required_inputs}. Usually
generated by \code{\link[=plugin_Read]{plugin_Read()}}.}
}
\value{
Actually launch the plugin UI.
}
\description{
Run Plugin UI
}
\keyword{internal}
