% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_dfAnalyticsInput}
\alias{sample_dfAnalyticsInput}
\title{Analytics Input Dataset}
\format{
A tibble with 2602 rows and 7 columns:
\describe{
\item{MetricID}{metric identifier}
\item{SubjectID}{unique subject identifier}
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Numerator}{numerator for this individual}
\item{Denominator}{denominator for this individual}
\item{Metric}{calculated rate/metric value}
}
}
\source{
Generated from data in the \code{clindata} package, using the \code{gsm}
package.
}
\usage{
sample_dfAnalyticsInput
}
\description{
Summary metric data for the "Site Subjects" module on the "Metric Details"
tab.
}
\keyword{datasets}
