% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_dfBounds}
\alias{sample_dfBounds}
\title{Metric Bounds Dataset}
\format{
A tibble with 2091 rows and 8 columns:
\describe{
\item{Threshold}{number of standard deviations that the upper and lower bounds are based on}
\item{Denominator}{calculated denominator value}
\item{LogDenominator}{calculated log denominator value}
\item{Numerator}{calculated numerator value}
\item{Metric}{calculated rate/metric value}
\item{MetricID}{unique metric identifier}
\item{StudyID}{unique study identifier}
\item{SnapshotDate}{date of the snapshot}
}
}
\source{
Generated from data in the \code{clindata} package, using the \code{gsm}
package.
}
\usage{
sample_dfBounds
}
\description{
Information about the bounds used to flag metrics.
}
\keyword{datasets}
