% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\docType{data}
\name{sample_dfGroups}
\alias{sample_dfGroups}
\title{Group Metadata Dataset}
\format{
A tibble with 574 rows and
4 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Param}{Parameter of interest for the group}
\item{Value}{Value of specified Param}
}
}
\source{
Generated from data in the \code{clindata} package, using the
\code{gsm.mapping} package.
}
\usage{
sample_dfGroups
}
\description{
Information about sites, countries, and the overall study in the sample data.
}
\seealso{
Other sample data: 
\code{\link{sample_dfAnalyticsInput}},
\code{\link{sample_dfBounds}},
\code{\link{sample_dfMetrics}},
\code{\link{sample_dfResults}},
\code{\link{sample_fnCountData}()},
\code{\link{sample_fnFetchData}()}
}
\concept{sample data}
\keyword{datasets}
