% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\docType{data}
\name{sample_dfGroups}
\alias{sample_dfGroups}
\title{Group Metadata Dataset}
\format{
A tibble with 3903 rows and 4 columns:
\describe{
\item{GroupID}{grouping variable}
\item{GroupLevel}{level of grouping variable}
\item{Param}{Parameter of interest for the group}
\item{Value}{Value of specified Param}
}
}
\source{
Generated from data in the \code{clindata} package, using the \code{gsm}
package.
}
\usage{
sample_dfGroups
}
\description{
Information about sites, countries, and the overall study in the sample data.
}
\keyword{datasets}
