% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\docType{data}
\name{sample_dfMetrics}
\alias{sample_dfMetrics}
\title{Metric Metadata Dataset}
\format{
A tibble with 22 rows and
16 columns:
\describe{
\item{MetricID}{unique metric identifier}
\item{File}{yaml file for workflow}
\item{GroupLevel}{level of grouping variable}
\item{Abbreviation}{abbreviation for the metric}
\item{Metric}{name of the metric}
\item{Numerator}{data source for the numerator}
\item{Denominator}{data source for the denominator}
\item{Model}{model used to calculate metric}
\item{Score}{type of score reported}
\item{Type}{statistical outcome type}
\item{Threshold}{thresholds to be used for bounds and flags}
\item{nMinDenominator}{minimum denominator required to return a score and calculate a flag}
}
}
\source{
Generated from data in the \code{clindata} package, using the
\code{gsm.mapping} package.
}
\usage{
sample_dfMetrics
}
\description{
Information about the metrics included in the sample data.
}
\seealso{
Other sample data: 
\code{\link{sample_dfAnalyticsInput}},
\code{\link{sample_dfBounds}},
\code{\link{sample_dfGroups}},
\code{\link{sample_dfResults}},
\code{\link{sample_fnCountData}()},
\code{\link{sample_fnFetchData}()}
}
\concept{sample data}
\keyword{datasets}
