% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\name{sample_FetchParticipantData}
\alias{sample_FetchParticipantData}
\title{Fetch Data for a Participant}
\usage{
sample_FetchParticipantData(strSubjectID)
}
\arguments{
\item{strSubjectID}{\code{character} A \code{SubjectID} of an individual participant.}
}
\value{
A list with elements \code{metadata} (a list of participant metadata) and
\code{metric_data} (a named list, where the names are the available metric
summaries, and the values are data frames containing the details).
}
\description{
This is a sample function demonstrating the type of function the user will
supply to find additional data about a given participant.
}
\examples{
sample_FetchParticipantData("0001")
}
