% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SampleData.R
\name{sample_fnCountData}
\alias{sample_fnCountData}
\title{Count Data for a Domain}
\usage{
sample_fnCountData(
  strDomainID = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY"),
  strGroupID = NULL,
  strGroupLevel = "Site",
  strSubjectID = NULL,
  dSnapshotDate = NULL
)
}
\arguments{
\item{strDomainID}{\code{character} The domain data.frame to load.}

\item{strGroupID}{\code{character} A \code{GroupID} of an individual site or other
group within a study.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: \code{NULL}
and taken from \code{dfMetrics$GroupLevel} if available.}

\item{strSubjectID}{\code{character} A \code{SubjectID} of an individual participant.}

\item{dSnapshotDate}{\code{Date} The date of a data snapshot.}
}
\value{
An integer signifying the number of rows available for the specified
domain.
}
\description{
This is a sample function demonstrating the type of function the user will
supply to count rows of additional data available for a given domain.
}
\examples{
sample_fnCountData("AE")
sample_fnCountData("AE", strGroupID = "0X103")
sample_fnCountData("AE", strSubjectID = "1350")
}
\seealso{
Other sample data: 
\code{\link{sample_dfAnalyticsInput}},
\code{\link{sample_dfBounds}},
\code{\link{sample_dfGroups}},
\code{\link{sample_dfMetrics}},
\code{\link{sample_dfResults}},
\code{\link{sample_fnFetchData}()}
}
\concept{sample data}
