% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Define Server}
\usage{
server(
  input,
  output,
  session,
  dfResults,
  dfGroups,
  dfMetrics,
  dfBounds,
  snapshot
)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Define Server
}
