% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Define Server}
\usage{
server(
  input,
  output,
  session,
  dfResults,
  dfGroups,
  dfMetrics,
  dfBounds,
  dfAnalyticsInput
)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}
}
\description{
Define Server
}
