% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The namespace id for this module.}

\item{combinedData}{\code{data.frame} A combined data set with the required
metrics.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfSite}{\code{data.frame} Optional site-level metadata.}

\item{dfStudy}{\code{data.frame} Optional study-level metadata.}

\item{domain}{The type of data and subsequent metrics you would like to see.}

\item{input}{Shiny inputs}

\item{lAssessment}{\code{list} List of KRI assessments from \code{params} within
\code{KRIReport.Rmd}.}

\item{metric}{The reactive value provided by the metric input from \code{server}}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{site}{The reactive value provided by the site input from \code{server}}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}

\item{subset_col}{The id column that will be selected from the domain.}

\item{subset_val}{The specific value that the subset_col will be filtered by.}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
