% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{dfSite}{\code{data.frame} Optional site-level metadata.}

\item{dfStudy}{\code{data.frame} Optional study-level metadata.}

\item{fnFetchParticipantData}{\code{function} A function that takes a single
\code{strSubjectID} argument such as "0001", and returns a list with components
\code{metadata} and \code{metric_data}. \code{metadata} should be a named charact vector
(or something that can be coerced to a character vector) of information
about the participant, and \code{metric_data} should be a named list of
data.frames, each of which contains information related to the named
metric.}

\item{lAssessment}{\code{list} List of KRI assessments from \code{params} within
\code{KRIReport.Rmd}.}

\item{rctv_strMetricID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the selected \code{MetricID}.}

\item{rctv_strSiteID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{strMetricID}{\code{string} A \code{MetricID} to focus on.}

\item{strSite}{\code{string} The \code{GroupID} of a site to focus on.}

\item{strSubjectID}{\code{string} A \code{SubjectID} of an individual participant.}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
