% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfSite}{\code{data.frame} Optional site-level metadata.}

\item{dfStudy}{\code{data.frame} Optional study-level metadata.}

\item{input}{Shiny inputs}

\item{lAssessment}{\code{list} List of KRI assessments from \code{params} within
\code{KRIReport.Rmd}.}

\item{output}{Shiny outputs}

\item{session}{Shiny session}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
