% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The id for this module or widget in this namespace.}

\item{chrLinkIDs}{\code{character} Module ids for multiple modules.}

\item{chrLabels}{\code{character} A vector of labels for display to the user.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}

\item{chrValues}{\code{character} A vector of values to associate with a vector of
labels.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{fnFetchParticipantData}{\code{function} A function that takes a single
\code{strSubjectID} argument such as "0001", and returns a list with components
\code{metadata} and \code{metric_data}. \code{metadata} should be a named charact vector
(or something that can be coerced to a character vector) of information
about the participant, and \code{metric_data} should be a named list of
data.frames, each of which contains information related to the named
metric.}

\item{intNParticipants}{\code{integer} The number of unique participants
(subjects) in the study.}

\item{lMetric}{\code{list} Named list of data describing a single metric, as well
as things like which group is selected.}

\item{lParticipantMetadata}{\code{list} Named list of data describing a single
participant.}

\item{lStudy}{\code{list} Named list of data describing the overall study.}

\item{rctv_lData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns a
named list of dataframes.}

\item{rctv_lParticipantMetricData}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}}
object that returns a list of dataframes for a given participant, one for
each metric that is available.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strName}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the name of an object, such as a particular dataframe in a named
list.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{strContainerID}{\code{character} The (namespaced) ID of the target container
(usually a div).}

\item{strID}{\code{character} An ID for this element.}

\item{strInputID}{\code{character} An ID to use for the Shiny input created by
this module or used by this Javascript.}

\item{strMetricID}{\code{character} A \code{MetricID} to focus on.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}

\item{strSubjectID}{\code{character} A \code{SubjectID} of an individual participant.}

\item{strText}{\code{character} Text to display.}

\item{strTitle}{\code{character} A title to display for the overall app.}

\item{strValue}{\code{character} The value of a field.}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
