% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{chrAllowedFields}{\code{character} A vector of names of allowed fields in an
object.}

\item{chrAllowedValues}{\code{character} A vector of allowed values for a vector.}

\item{chrDomains}{\code{character} A vector of names of domain data.frames to show
in the app. Supported domains are shown in the default value.}

\item{chrLinkIDs}{\code{character} Module ids for multiple modules.}

\item{chrLabels}{\code{character} A vector of labels for display to the user.}

\item{chrMessage}{\code{character} A vector of message elements to be formatted
via \code{\link[cli:cli_bullets]{cli::cli_bullets()}}.}

\item{chrMetrics}{\code{character} A named vector of MetricIDs, where the names
are the full Metric.}

\item{chrPluginFiles}{\code{character} The files in a plugin directory.}

\item{chrRequiredColumns}{\code{character} A vector of expected columns in a
data.frame.}

\item{chrRequiredFields}{\code{character} A vector of names of required fields in
an object.}

\item{chrSites}{\code{character} A vector of sites available in the study.}

\item{chrValues}{\code{character} A vector of values to associate with a vector of
labels.}

\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfParticipantGroups}{\code{data.frame} Unique \code{SubjectID} and \code{GroupID}
combos from \code{dfAnalyticsInput}.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}

\item{envEvaluate}{\code{environment} The environment in which any variables used
in the message are defined. You almost definitely want to leave this as the
default value.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomain} argument
and optional \code{strSiteID} and \code{strSubjectID} and returns a data.frame. See
\code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The returned data.frame contains
information about the named domain.}

\item{fnServer}{\code{function} A Shiny server function that takes arguments
\code{input}, \code{output}, and \code{session}. This function will be called at the start
of the main app server function.}

\item{id}{\code{character} The id for this element.}

\item{intKRIColorCount}{\code{integer} A named vector of counts by color.}

\item{intNParticipants}{\code{integer} The number of unique participants
(subjects) in the study.}

\item{intRed}{\code{integer} The number of sites with at least one red flag.}

\item{intAmber}{\code{integer} The number of sites with at least one amber flag.}

\item{lMetric}{\code{list} Named list of data describing a single metric, as well
as things like which group is selected.}

\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}

\item{lPluginDefinition}{\code{list} A named list with required elements
\code{strTitle}, \code{fnUI}, and \code{fnServer}, and optional fields \code{packages} and
\code{lConfig}. Usually generated by \code{\link[=plugin_Read]{plugin_Read()}}.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{lStudy}{\code{list} Named list of data describing the overall study.}

\item{rctv_dfBounds}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a set of predicted percentages/rates and upper- and lower-bounds
across the full range of sample sizes/total exposure values for reporting.}

\item{rctv_dfData}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a generic dataframe.}

\item{rctv_dfDomain}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a domain dataframe.}

\item{rctv_dfResults}{\verb{reactive dataframe} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a stacked summary of analysis pipeline output.}

\item{rctv_gtObject}{\verb{reactive gt_table} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns a \code{\link[gt:gt]{gt::gt()}} object.}

\item{rctv_lMetric}{\verb{reactive list} A \code{\link[shiny:reactive]{shiny::reactive()}} object that returns
a named list of data describing a single metric, as well as things like
which group is selected.}

\item{rctv_lglState}{\verb{reactive logical} A \code{\link[shiny:reactive]{shiny::reactive()}}] object that
returns a Boolean value indicating whether something is "off" (\code{FALSE}) or
"on" (\code{TRUE}).}

\item{rctv_strDomainID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{DomainID} (such as "AE" or "SUBJ").}

\item{rctv_strCurrentTab}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the currently selected tab.}

\item{rctv_strGroupSubset}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected subset of groups to include in the table.}

\item{rctv_strInput}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value(s) of an input to share with an interactive gt table. If
this argument is a \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}, it will be used to push updates
back to the calling function.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strName}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the name of an object, such as a particular dataframe in a named
list.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{rctv_strValue}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the value of a field.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{strArg}{\code{character} The name of the argument in the calling function.
In general, this value should either be left as the default, or passed from
the calling function if the calling function also has a \code{strArg} argument.}

\item{strClass}{\code{character} A descriptive label for this type of error, in
lower_snake_case.}

\item{strColorFamily}{\code{character} Whether to load the \code{"dark"} version of
this color, or the \code{"light"} version. Default: \code{"dark"}.}

\item{strColorName}{\code{character} The target color to report about.}

\item{strColorCode}{\code{character} The hex code (such as \code{"#FFFFFF"}) for a
color.}

\item{strContainerID}{\code{character} The (namespaced) ID of the target container
(usually a div).}

\item{strDomain}{\code{character} The domain data.frame to load. Supported domains
are shown in the default value.}

\item{strEmpty}{\code{character} The value to return when everything is
deselected.}

\item{strFavicon}{\code{character} The name of an icon to use in the browser tab
via \code{\link[favawesome:fav]{favawesome::fav()}}.}

\item{strFaviconColor}{\code{character} The hexcode or name of a color to use as
the icon fill for \code{\link[favawesome:fav]{favawesome::fav()}}.}

\item{strFunction}{\code{character} The name of a function, with or without
namespace ("reactive" or "shiny::reactive"). If a namespace is supplied, we
check that the package is installed, and attach it. The function can also
be supplied as a function or a purrr-style formula. See
\code{\link[rlang:as_function]{rlang::as_function()}} (the \code{x} argument) for more details.}

\item{strGroupID}{\code{character} A \code{GroupID} to focus on.}

\item{strGroupLabelKey}{\code{character} Value for the group label key. Default:
\code{"InvestigatorLastName"}.}

\item{strGroupLevel}{\code{character} Value for the group level. Default: \code{NULL}
and taken from \code{dfMetrics$GroupLevel} if available.}

\item{strGroupSubset}{\code{character} Subset of groups to include in the table.
Default: \code{"red"}. Options:
\itemize{
\item \code{"all"}: All groups.
\item \code{"red"}: Groups with 1+ red flags.
\item \code{"red/amber"}: Groups with 1+ red/amber flag.
\item \code{"amber"}: Groups with 1+ amber flag.
}}

\item{strInputID}{\code{character} An ID to use for the Shiny input created by
this module or used by this JavaScript.}

\item{strLabel}{\code{character} The label of a field.}

\item{strMetricID}{\code{character} A \code{MetricID} to focus on.}

\item{strOutcome}{\code{character} Outcome variable. Default: \code{"Score"}.}

\item{strPlotTitle}{\code{character} A title for a plot, usually the name of a
metric.}

\item{strSiteID}{\code{character} A \code{GroupID} of an individual site within a
study.}

\item{strSubjectID}{\code{character} A \code{SubjectID} of an individual participant.}

\item{strTargetTab}{\code{character} The tab to switch to.}

\item{strText}{\code{character} Text to display.}

\item{strTitle}{\code{character} A title to display for the overall app.}

\item{strValue}{\code{character} The value of a field.}

\item{strWhat}{\code{character} A sentence-case description of the object being
inspected.}

\item{tagListSidebar}{\code{taglist} An optional \code{\link[htmltools:tagList]{htmltools::tagList()}} of
additional elements to add to the top of the app sidebar.}

\item{x}{An object to validate.}
}
\description{
Reused parameter definitions are gathered here for easier usage. Note: We
intentionally use \code{id} rather than the more standardized \code{strID} throughout
this package, because \code{\link[shiny:testServer]{shiny::testServer()}} specifically expects an \code{id}
argument in module server functions.
}
\keyword{internal}
