% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_panel_meta_tag_list.R
\name{side_panel_meta_tag_list}
\alias{side_panel_meta_tag_list}
\title{Meta Tag List for Side Panel}
\usage{
side_panel_meta_tag_list(snapshot)
}
\arguments{
\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Meta Tag List for Side Panel
}
