% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_details_metric_meta_data_list.R
\name{site_details_metric_meta_data_list}
\alias{site_details_metric_meta_data_list}
\title{Creates taglist to summarize participants at a particular site}
\usage{
site_details_metric_meta_data_list(dfMetrics, rctv_strMetricID)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{rctv_strMetricID}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the selected \code{MetricID}.}
}
\description{
Creates taglist to summarize participants at a particular site
}
\keyword{internal}
