% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_site_details_server.R
\name{site_details_server}
\alias{site_details_server}
\title{Site Details Server}
\usage{
site_details_server(id, snapshot, site)
}
\arguments{
\item{id}{The namespace id}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}

\item{site}{The reactive value provided by the site input from \code{server}}
}
\description{
Site Details Server
}
