% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dfDomains.R
\name{srvr_dfDomain}
\alias{srvr_dfDomain}
\title{Reactive domain df}
\usage{
srvr_dfDomain(strDomainID, fnFetchData, rctv_strSiteID, rctv_strSubjectID)
}
\arguments{
\item{strDomainID}{\code{character} The domain data.frame to load.}

\item{fnFetchData}{\code{function} A function that takes a \code{strDomainID} argument
and optional \code{strSiteID}, \code{strSubjectID}, and/or \code{dSnapshotDate}, and
returns a data.frame. See \code{\link[=sample_fnFetchData]{sample_fnFetchData()}} for an example. The
returned data.frame contains information about the named domain. If the
function throws an error, the error is elevated to the user, so you can use
errors to pass requirements through to the user.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} that generates a domain df, or \code{NULL}. If
\code{fnFetchData()} throws an error, that error is passed to the user via
\code{\link[=srvr_ShowConditionMessage]{srvr_ShowConditionMessage()}}.
}
\description{
Reactive domain df
}
\keyword{internal}
