% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_PluginHelpers.R
\name{srvr_IngestReactive}
\alias{srvr_IngestReactive}
\title{Ingest Data Reactively}
\usage{
srvr_IngestReactive(
  l_rctvDomains,
  lSpec,
  l_rctvDomainHashes,
  rctv_bPluginReady
)
}
\arguments{
\item{l_rctvDomains}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects, each
of which returns a domain dataframe.}

\item{l_rctvDomainHashes}{\code{list} A named list of \code{\link[shiny:reactive]{shiny::reactive()}} objects,
each of which returns the \code{\link[rlang:hash]{rlang::hash()}} of a domain dataframe.}
}
\value{
A named list of ingested data.
}
\description{
Ingest Data Reactively
}
\keyword{internal}
