% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_MetricDetails.R
\name{srvr_MetricDetails}
\alias{srvr_MetricDetails}
\title{Temporary Metric Details wrapper}
\usage{
srvr_MetricDetails(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  rctv_strMetricID,
  rctv_strPrimaryNavBar,
  rctv_strSiteID,
  rctv_strSubjectID,
  input,
  output,
  session
)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}

\item{rctv_strSubjectID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the \code{SubjectID} of the selected participant.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} for \code{rctv_strSiteDetailsParticipant} from
\code{\link[=mod_SiteDetails_Server]{mod_SiteDetails_Server()}}.
}
\description{
This really should be a full module, but this lets me "hide" some server code
so it isn't all inside \code{\link[=gsmApp_Server]{gsmApp_Server()}}.
}
\keyword{internal}
