% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_rctv_chrParticipantIDs}
\alias{srvr_rctv_chrParticipantIDs}
\title{Reactive participant IDs}
\usage{
srvr_rctv_chrParticipantIDs(dfAnalyticsInput, rctv_strSiteID)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} object
that returns the \code{GroupID} of the selected site, and can be used to update
the selected site.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} of dfMetrics as a list, filtered by metric.
}
\description{
Reactive participant IDs
}
\keyword{internal}
