% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_rctv_chrParticipantIDs}
\alias{srvr_rctv_chrParticipantIDs}
\title{Reactive participant IDs}
\usage{
srvr_rctv_chrParticipantIDs(dfParticipantGroups, rctv_strSiteID)
}
\arguments{
\item{dfParticipantGroups}{\code{data.frame} Unique \code{SubjectID} and \code{GroupID}
combos from \code{dfAnalyticsInput}.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} of dfMetrics as a list, filtered by metric.
}
\description{
Reactive participant IDs
}
\keyword{internal}
