% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_rctv_lMetric}
\alias{srvr_rctv_lMetric}
\title{Reactive lMetric by metric and site}
\usage{
srvr_rctv_lMetric(
  dfMetrics,
  rctv_strMetricID,
  rctv_strSiteID,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{rctv_strSiteID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} of dfMetrics as a list, with information about
the selected site.
}
\description{
Reactive lMetric by metric and site
}
\keyword{internal}
