% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SharedReactives.R
\name{srvr_rctv_lMetric_base}
\alias{srvr_rctv_lMetric_base}
\title{Reactive lMetric by metric}
\usage{
srvr_rctv_lMetric_base(
  dfMetrics,
  rctv_strMetricID,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{rctv_strMetricID}{\verb{reactive character} A \code{\link[shiny:reactive]{shiny::reactive()}} object
that returns the selected \code{MetricID}.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} of dfMetrics as a list, filtered by metric.
}
\description{
Reactive lMetric by metric
}
\keyword{internal}
