% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_Reset.R
\name{srvr_Reset}
\alias{srvr_Reset}
\title{Reload app from reset button}
\usage{
srvr_Reset(input, session)
}
\arguments{
\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} to apply the reset.
}
\description{
Reload app from reset button
}
\keyword{internal}
