% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvr_SyncInput.R
\name{srvr_SyncVirtualSelectInput}
\alias{srvr_SyncVirtualSelectInput}
\title{Synchronize virtualSelectInput with reactive}
\usage{
srvr_SyncVirtualSelectInput(
  id,
  session = getDefaultReactiveDomain(),
  rctv_chrChoices = reactive(NULL, domain = session),
  rctv_strLabel = reactive(NULL, domain = session),
  rctv_strSelected = reactive(NULL, domain = session)
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{session}{\code{environment} Session from which to make a child scope (the
default should almost always be used).}

\item{rctv_chrChoices}{\verb{reactive character} Choices to set.}

\item{rctv_strLabel}{\verb{reactive character} Label to set.}

\item{rctv_strSelected}{\verb{reactive character} Selected value to set.}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} that triggers
\code{\link[shinyWidgets:updateVirtualSelect]{shinyWidgets::updateVirtualSelect()}} when the connected reactive changes.
}
\description{
Synchronize virtualSelectInput with reactive
}
\keyword{internal}
