% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_study_overview_server.R
\name{study_overview_server}
\alias{study_overview_server}
\title{Study Overview Server}
\usage{
study_overview_server(id, dfResults, dfMetrics, dfGroups)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{dfMetrics}{\code{data.frame} Metric-specific metadata for use in charts and
reporting.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}
}
\description{
Study Overview Server
}
