% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_study_overview_server.R
\name{study_overview_server}
\alias{study_overview_server}
\title{Study Overview Server}
\usage{
study_overview_server(id, snapshot)
}
\arguments{
\item{id}{The namespace id for this module.}

\item{snapshot}{A \code{list} object containing the data for a snapshot. This
object funnels into virtually every function of the application.}
}
\description{
Study Overview Server
}
