% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_participants_by_site.R
\name{table_participants_by_site}
\alias{table_participants_by_site}
\title{Table of Participants Based on Site Selection}
\usage{
table_participants_by_site(dfAnalyticsInput, rctv_strSite)
}
\arguments{
\item{dfAnalyticsInput}{\code{data.frame} Participant-level metric data.}

\item{rctv_strSite}{\verb{reactive string} A \code{\link[shiny:reactive]{shiny::reactive()}} object that
returns the \code{GroupID} of a site.}
}
\description{
Table of Participants Based on Site Selection
}
