% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_participant_select.R
\name{update_participant_select}
\alias{update_participant_select}
\title{Update Participant Select}
\usage{
update_participant_select(input, output, session, snapshot, site)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{Shiny session}

\item{snapshot}{The snapshot \code{list} object passed from \code{run_app()}}

\item{site}{The user-selected site.}
}
\description{
Update Participant Select
}
