% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Validate.R
\name{validate_chrDomains}
\alias{validate_chrDomains}
\title{Confirm that an object is valid chrDomains}
\usage{
validate_chrDomains(chrDomains, lPlugins = NULL, envCall = rlang::caller_env())
}
\arguments{
\item{chrDomains}{\code{character} A (named) vector of domains to include in the
app. The values of the vector will be used as labels, and the names will be
used as IDs. The IDs will be passed to \code{fnFetchData()} to fetch data about
that domain.}

\item{lPlugins}{\code{list} Optional list of plugins to include in the app.}

\item{envCall}{\code{environment} The environment from which this function was
called, for use in better error messages. This value should usually be left
as the default, or passed from the calling function if the calling function
also has an \code{envCall} argument.}
}
\value{
\code{chrDomains} with names in all caps.
}
\description{
Confirm that an object is valid chrDomains
}
\keyword{internal}
