% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_BarChart.R
\name{Widget_BarChart}
\alias{Widget_BarChart}
\title{Bar Chart Widget}
\usage{
Widget_BarChart(
  id,
  dfResults,
  strOutcome = "Score",
  lMetric = list(),
  dfGroups = NULL,
  dfBounds = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this module, widget, or other element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.
This will be filtered to cases where \code{GroupLevel == "Site"}.}

\item{strOutcome}{\code{character} Outcome variable. Default: \code{"Score"}.}

\item{lMetric}{\code{list} Named list of data describing a single metric, as well
as things like which group is selected.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A widget that generates a bar chart of group-level metric results,
plotting the denominator on the x-axis and the numerator on the y-axis.
}
\keyword{internal}
