% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TimeSeries.R
\name{Widget_TimeSeries}
\alias{Widget_TimeSeries}
\title{Time Series Widget}
\usage{
Widget_TimeSeries(
  id,
  dfResults,
  strOutcome = "Score",
  lMetric = list(),
  dfGroups = NULL,
  dfBounds = NULL
)
}
\arguments{
\item{id}{\code{character} The id for this element.}

\item{dfResults}{\code{data.frame} A stacked summary of analysis pipeline output.}

\item{strOutcome}{\code{character} Outcome variable. Default: \code{"Score"}.}

\item{lMetric}{\code{list} Named list of data describing a single metric, as well
as things like which group is selected.}

\item{dfGroups}{\code{data.frame} Group-level metadata dictionary.}

\item{dfBounds}{\code{data.frame} Set of predicted percentages/rates and upper-
and lower-bounds across the full range of sample sizes/total exposure
values for reporting.}
}
\description{
Time Series Widget
}
\keyword{internal}
